% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag.R
\name{access_dag}
\alias{access_dag}
\alias{edges,dag-method}
\alias{edges<-,dag-method}
\alias{distances,dag-method}
\alias{distances<-,dag-method}
\alias{distance_units,dag-method}
\alias{distance_units<-,dag-method}
\title{Get or set slots from an object of class \code{dag}.}
\usage{
\S4method{edges}{dag}(x)

\S4method{edges}{dag}(x) <- value

\S4method{distances}{dag}(x)

\S4method{distances}{dag}(x) <- value

\S4method{distance_units}{dag}(x)

\S4method{distance_units}{dag}(x) <- value
}
\arguments{
\item{x}{An object of class \code{dag}.}

\item{value}{A replacement value.}
}
\description{
Get or set slots from an object of class \code{dag}.
}
\concept{access_dag}
