% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_staRVe.R
\name{Access_staRVe}
\alias{Access_staRVe}
\alias{observation}
\alias{observation<-}
\alias{process}
\alias{process<-}
\alias{parameters}
\alias{parameters<-}
\alias{convergence}
\alias{convergence<-}
\alias{settings}
\alias{settings<-}
\title{Get or set slots from an object of class \code{staRVe}.}
\usage{
observation(x)

observation(x) <- value

process(x)

process(x) <- value

parameters(x)

parameters(x) <- value

convergence(x)

convergence(x) <- value

settings(x)

settings(x) <- value
}
\arguments{
\item{x}{An object of class \code{staRVe}.}

\item{value}{A replacement value.}
}
\description{
Get or set slots from an object of class \code{staRVe}.
}
\concept{Access_staRVe}
