% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/staRVe_fit.R
\name{TMB_out}
\alias{TMB_out}
\alias{TMB_out<-}
\alias{tracing}
\alias{tracing<-}
\alias{Access_staRVe_fit}
\title{Get or set slots from an object of class \code{staRVe_fit}.}
\usage{
TMB_out(x)

TMB_out(x) <- value

tracing(x)

tracing(x) <- value
}
\arguments{
\item{x}{An object of class \code{staRVe_fit}.}

\item{value}{A replacement value}
}
\description{
Get or set slots from an object of class \code{staRVe_fit}.
}
\concept{Access_staRVe_fit}
