% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/staRVe_model.R
\name{graph}
\alias{graph}
\alias{observations}
\alias{observations<-}
\alias{process}
\alias{process<-}
\alias{settings}
\alias{settings<-}
\alias{Access_staRVe_model}
\title{Get or set slots from an object of class \code{staRVe_model}.}
\usage{
graph(x)

observations(x)

observations(x) <- value

process(x)

process(x) <- value

settings(x)

settings(x) <- value
}
\arguments{
\item{x}{An object of class \code{staRVe_model}.}

\item{value}{A replacement value}
}
\description{
Get or set slots from an object of class \code{staRVe_model}.
}
\concept{Access_staRVe_model}
