% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/staRVe_observation_parameters.R
\name{fixed_effects}
\alias{fixed_effects}
\alias{fixed_effects<-}
\alias{link_function}
\alias{link_function<-}
\alias{response_distribution}
\alias{response_distribution<-}
\alias{response_parameters}
\alias{response_parameters<-}
\alias{Access_staRVe_observation_parameters}
\title{Get or set slots from an object of class \code{staRVe_observation_parameters}.}
\usage{
fixed_effects(x)

fixed_effects(x) <- value

link_function(x)

link_function(x) <- value

response_distribution(x)

response_distribution(x) <- value

response_parameters(x)

response_parameters(x) <- value
}
\arguments{
\item{x}{An object of class \code{staRVe_observation_parameters}.}

\item{value}{A replacement value}
}
\description{
Get or set slots from an object of class \code{staRVe_observation_parameters}.
}
\concept{Access_staRVe_observation_parameters}
