% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staRVe_observations.R
\name{access_staRVe_observations}
\alias{access_staRVe_observations}
\alias{data,staRVe_observations-method}
\alias{data<-,staRVe_observations-method}
\alias{transient_graph,staRVe_observations-method}
\alias{transient_graph<-,staRVe_observations-method}
\alias{parameters,staRVe_observations-method}
\alias{parameters<-,staRVe_observations,ANY-method}
\title{Get or set slots from an object of class \code{staRVe_observations}.}
\usage{
\S4method{data}{staRVe_observations}(x)

\S4method{data}{staRVe_observations}(x) <- value

\S4method{transient_graph}{staRVe_observations}(x)

\S4method{transient_graph}{staRVe_observations}(x) <- value

\S4method{parameters}{staRVe_observations}(x)

\S4method{parameters}{staRVe_observations,ANY}(x) <- value
}
\arguments{
\item{x}{An object of class \code{staRVe_observations}.}

\item{value}{A replacement value}
}
\description{
Get or set slots from an object of class \code{staRVe_observations}.
}
\concept{access_staRVe_observations}
