% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/staRVe_process.R
\name{parameters}
\alias{parameters}
\alias{parameters<-}
\alias{persistent_graph}
\alias{persistent_graph<-}
\alias{random_effects}
\alias{random_effects<-}
\alias{Access_staRVe_process}
\title{Get or set slots from an object of class \code{staRVe_process}.}
\usage{
parameters(x)

parameters(x) <- value

persistent_graph(x)

persistent_graph(x) <- value

random_effects(x)

random_effects(x) <- value
}
\arguments{
\item{x}{An object of class \code{staRVe_process}.}

\item{value}{A replacement value}
}
\description{
Get or set slots from an object of class \code{staRVe_process}.
}
\concept{Access_staRVe_process}
