% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staRVe_process_parameters.R
\name{access_staRVe_process_parameters}
\alias{access_staRVe_process_parameters}
\alias{covariance_function,staRVe_process_parameters-method}
\alias{covariance_function<-,staRVe_process_parameters-method}
\alias{spatial_parameters,staRVe_process_parameters-method}
\alias{spatial_parameters<-,staRVe_process_parameters-method}
\alias{time_parameters,staRVe_process_parameters-method}
\alias{time_parameters<-,staRVe_process_parameters-method}
\title{Get or set slots from an object of class \code{staRVe_process_parameters}.}
\usage{
\S4method{covariance_function}{staRVe_process_parameters}(x)

\S4method{covariance_function}{staRVe_process_parameters}(x) <- value

\S4method{spatial_parameters}{staRVe_process_parameters}(x)

\S4method{spatial_parameters}{staRVe_process_parameters}(x) <- value

\S4method{time_parameters}{staRVe_process_parameters}(x)

\S4method{time_parameters}{staRVe_process_parameters}(x) <- value
}
\arguments{
\item{x}{An object of class \code{staRVe_process_parameters}.}

\item{value}{A replacement value}
}
\description{
Get or set slots from an object of class \code{staRVe_process_parameters}.
}
\concept{access_staRVe_process_parameters}
