% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/staRVe_settings.R
\name{distance_units}
\alias{distance_units}
\alias{distance_units<-}
\alias{formula}
\alias{formula<-}
\alias{max_distance}
\alias{max_distance<-}
\alias{n_neighbours}
\alias{n_neighbours<-}
\alias{p_far_neighbours}
\alias{p_far_neighbours<-}
\alias{Access_staRVe_settings}
\title{Get or set slots from an object of class \code{staRVe_settings}.}
\usage{
distance_units(x)

distance_units(x) <- value

formula(x, ...)

formula(x) <- value

max_distance(x)

max_distance(x) <- value

n_neighbours(x)

n_neighbours(x) <- value

p_far_neighbours(x)

p_far_neighbours(x) <- value
}
\arguments{
\item{x}{An object of class \code{staRVe_settings}.}

\item{value}{A replacement value}
}
\description{
Get or set slots from an object of class \code{staRVe_settings}.
}
\concept{Access_staRVe_settings}
