% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staRVe_tracing.R
\name{access_staRVe_tracing}
\alias{access_staRVe_tracing}
\alias{opt_time,staRVe_tracing-method}
\alias{opt_time<-,staRVe_tracing-method}
\alias{hess_time,staRVe_tracing-method}
\alias{hess_time<-,staRVe_tracing-method}
\alias{sdr_time,staRVe_tracing-method}
\alias{sdr_time<-,staRVe_tracing-method}
\alias{parameter_hessian,staRVe_tracing-method}
\alias{parameter_hessian<-,staRVe_tracing-method}
\alias{parameter_covariance,staRVe_tracing-method}
\alias{parameter_covariance<-,staRVe_tracing-method}
\alias{timing,staRVe_tracing-method}
\title{Get or set slots from an object of class \code{staRVe_tracing}.}
\usage{
\S4method{opt_time}{staRVe_tracing}(x)

\S4method{opt_time}{staRVe_tracing}(x) <- value

\S4method{hess_time}{staRVe_tracing}(x)

\S4method{hess_time}{staRVe_tracing}(x) <- value

\S4method{sdr_time}{staRVe_tracing}(x)

\S4method{sdr_time}{staRVe_tracing}(x) <- value

\S4method{parameter_hessian}{staRVe_tracing}(x)

\S4method{parameter_hessian}{staRVe_tracing}(x) <- value

\S4method{parameter_covariance}{staRVe_tracing}(x)

\S4method{parameter_covariance}{staRVe_tracing}(x) <- value

\S4method{timing}{staRVe_tracing}(x)
}
\arguments{
\item{x}{An object of class \code{staRVe_tracing}.}

\item{value}{A replacement value}
}
\description{
Get or set slots from an object of class \code{staRVe_tracing}.
}
\concept{access_staRVe_tracing}
