% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/TMB_out.R
\name{get_geo_vars}
\alias{get_geo_vars}
\alias{get_obs}
\alias{get_observation}
\alias{get_parameters}
\alias{get_pars}
\alias{get_proc}
\alias{get_process}
\alias{get_resp}
\alias{get_working_pars}
\alias{obj}
\alias{obj<-}
\alias{opt}
\alias{opt<-}
\alias{report}
\alias{sdr}
\alias{sdr<-}
\alias{sdreport}
\alias{symbolicAnalysis}
\alias{symbolicAnalysis<-}
\alias{TMB_in}
\alias{Access_TMB_out}
\title{Get or set slots from an object of class \code{TMB_out}.}
\usage{
get_geo_vars(x, var, sf_obj, get_sd)

get_obs(x, sf_obj)

get_observation(x, sf_obj)

get_parameters(x)

get_pars(x)

get_proc(x, sf_obj)

get_process(x, sf_obj)

get_resp(x, sf_obj)

get_working_pars(x)

obj(x)

obj(x) <- value

opt(x)

opt(x) <- value

report(x)

sdr(x)

sdr(x) <- value

sdreport(x)

symbolicAnalysis(x)

symbolicAnalysis(x) <- value

TMB_in(x)
}
\arguments{
\item{x}{An object of class \code{TMB_out}.}

\item{value}{A replacement value}
}
\description{
Get or set slots from an object of class \code{TMB_out}.
}
\concept{Access_TMB_out}
