% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag.R
\name{Cdag_to_Rdag}
\alias{Cdag_to_Rdag}
\title{Increase all the parent indices of a directed acyclic graph by one for use in R.}
\usage{
Cdag_to_Rdag(dag)
}
\arguments{
\item{dag}{A directed acyclic graph. See \code{\link{construct_dag}}.}
}
\value{
The same directed acyclic graph as passed in with all parent indices
 increased by one.
}
\description{
Increase all the parent indices of a directed acyclic graph by one for use in R.
}
