% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag.R
\name{construct_dag}
\alias{construct_dag}
\title{Construct a directed acyclic graph for an \code{sf} object with point geometry.}
\usage{
construct_dag(
  x,
  n_neighbours = 10,
  p_far_neighbours = 0.2,
  max_dist = Inf,
  distance_units = "km",
  silent = T
)
}
\arguments{
\item{x}{An \code{sf} object containing point geometries, which will be the
nodes of the graph.}

\item{n_neighbours}{A positive integer giving the (maximum) number of
parents for each node.}

\item{p_far_neighbours}{What percent of neighbours should be randomly selected?}

\item{max_dist}{The maximum distance (in meters) to search for parents.}

\item{distance_units}{Which units should be used for distances?}

\item{silent}{Should intermediate calculations be shown?}
}
\description{
The directed acyclic graph is stored as an ordered list. The i'th element of
 the list is a vector containing indices j such that there is a directed edge
 from node j to node i. In other words, each node j is a directed parent of node i.
}
\details{
There can only be an edge from vertex i to vertex j if i comes before j in
 the rows of \code{x}.
}
