% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/process.R
\name{create_graph_idx}
\alias{create_graph_idx}
\alias{create_graph_idx,sf,process-method}
\title{Create an index connecting locations to a graph}
\usage{
create_graph_idx(x, y, ...)

\S4method{create_graph_idx}{sf,process}(x, y, settings)
}
\arguments{
\item{x}{An sf object}

\item{y}{A process object}
}
\value{
An integer vector with size equal to the number of rows of x.
  The location of row i of x will be the same location as locations
  answer[i] of the graph of y.
}
\description{
Create an index connecting locations to a graph
}
\section{Methods (by class)}{
\itemize{
\item \code{x = sf,y = process}: Locations in the sf object x will be given an
index pointing to either a persistent graph or transient graph location of y.
}}

\keyword{internal}
