% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{dag}
\alias{dag}
\title{An S4 class to hold a directed acyclic graph with distances.}
\description{
An S4 class to hold a directed acyclic graph with distances.
}
\section{Slots}{

\describe{
\item{\code{edges}}{A list of integer vector, indexing the parents of each node.}

\item{\code{distances}}{A list of matrices containing the distances between
nodes and parents.}

\item{\code{distance_units}}{The units used for distance calculation.  Must be compatible
with units::set_units.}
}}

