% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{.covariance_from_formula}
\alias{.covariance_from_formula}
\title{Retrieve a spatial covariance function from a formula.}
\usage{
.covariance_from_formula(x, data = data.frame(1))
}
\arguments{
\item{x}{A formula object with a space(covariance,nu) function. The covariance
argument must be one those listed in get_staRVe_distributions("covariance").
The `nu' argument is only necessary when using the Matern covariance, and
supplying it fixes the value of nu (the smoothness parameter).}
}
\value{
A list containing the name of the covariance function and the value of nu.
}
\description{
Retrieve a spatial covariance function from a formula.
}
