% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.distances_to_weights}
\alias{.distances_to_weights}
\title{Convert an array of distances from random effects to observations into a weight matrix}
\usage{
.distances_to_weights(
  d,
  invarfun = base::min,
  betweenvarfun = base::max,
  weightfun = base::Vectorize(function(x) return(1)),
  normalize = FALSE,
  ...
)
}
\arguments{
\item{d}{A list of arrays. Each entry in the list must correspond to a single year. Each array is of dimension [# random effects,# observations,# response variables] where entry i,j,k is the distance from random effect i to observation j of response variable k. If variable k has no observation at j then the corresponding entry should be Inf.}

\item{invarfun}{A function to summarize the distance from each random effect to the observations of a single response variable.}

\item{betweenvarfun}{A function to summarize the distance from each random effect to the summarized distances (output of invarfun) of all the response variables.}

\item{weightfun}{A function that takes a distance and outputs a weight. The range of this function should be between 0 and 1, where typically f(0) = 1 and f(Inf) = 0.}

\item{normalize}{Logical. Should the weights have mean 1?}
}
\value{
A matrix with dimensions [# random effects, # times] where entry i,j is the weight for random effect i at time j.
}
\description{
Convert an array of distances from random effects to observations into a weight matrix
}
