% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag.R
\name{.get_one_dag_node}
\alias{.get_one_dag_node}
\title{Get parent nodes for a node of a directed acyclic graph.}
\usage{
.get_one_dag_node(x, nodes, settings, silent = T, ...)
}
\arguments{
\item{x}{An \code{sf} object containing one point geometry (if more than one
is present, only the first will be used).}

\item{nodes}{An \code{sf} object containing point geometries. These will
be the possible parent nodes of \code{x}.}

\item{settings}{An object of class \code{staRVe_settings}.}

\item{silent}{Should intermediate calculations be shown?}
}
\value{
A list with elements "parents" and "dists". The "parents" element
  is a vector giving the indices of the rows of \code{nodes} which are the
  parents of \code{x}. The "dists" element is a matrix whose first row is the
  distance between \code{x} and its parents, and the rest of the matrix gives
  the distances between parent nodes.
}
\description{
Get parent nodes for a node of a directed acyclic graph.
}
