% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{.mean_design_from_formula}
\alias{.mean_design_from_formula}
\title{Retrieve observation mean covariate data from a formula and a data.frame.}
\usage{
.mean_design_from_formula(x, data)
}
\arguments{
\item{x}{A formula object with terms grouped in a \code{mean(...)} function.}

\item{data}{A data.frame containing the covariate data.}
}
\value{
The design matrix for the covariates.
}
\description{
Retrieve observation mean covariate data from a formula and a data.frame.
}
