% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.names_from_formula}
\alias{.names_from_formula}
\title{Check which covariates are used in a staRVe formula}
\usage{
.names_from_formula(x)
}
\arguments{
\item{x}{A formula object with terms grouped in a \code{mean(...)} function.}
}
\value{
A character vector giving the names of covariates used in a formula.
}
\description{
Check which covariates are used in a staRVe formula
}
