% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{.predict_one_location}
\alias{.predict_one_location}
\title{Predict from a \code{staRVe} object at a single new location.}
\usage{
.predict_one_location(
  location,
  x,
  time,
  covariates,
  n_neighbours,
  max_dist,
  distance_units,
  parameter_covariance,
  ...
)
}
\arguments{
\item{location}{An object of class \code{sf} containing one point geometry.
There should also be a \code{parents} column containing an integer vector
entry, and a \code{dists} column containing a distance matrix.}

\item{x}{An object of class \code{staRVe}.}

\item{time}{The time at which to predict.}

\item{n_neighbours}{The maximum number of parents used to predict.}
}
\description{
Predict from a \code{staRVe} object at a single new location.
}
