% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.response_from_formula}
\alias{.response_from_formula}
\title{Retrieve response data from a formula and a data.frame.}
\usage{
.response_from_formula(x, data)
}
\arguments{
\item{x}{A formula object with a non-empty left-hand side.}

\item{data}{A data.frame to retrieve the response data from.}
}
\value{
A vector containing the response variable, with a `name' attribute
 containing the name of the response variable.
}
\description{
Retrieve response data from a formula and a data.frame.
}
