% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.sample_size_from_formula}
\alias{.sample_size_from_formula}
\title{Retrieve a sample size from a formula and a data.frame.}
\usage{
.sample_size_from_formula(x, data, nullReturn = F)
}
\arguments{
\item{x}{A formula object with a sample.size(...) function containing one entry.}

\item{data}{A data.frame containing the sample size information.}

\item{nullReturn}{If there is no sample.size(...) function, should a vector of ones be returned?}
}
\value{
A data.frame with a single column for the sample size.
}
\description{
Retrieve a sample size from a formula and a data.frame.
}
