% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.sf_from_raster_list}
\alias{.sf_from_raster_list}
\title{Reform a list of \code{Raster*} objects to an \code{sf} data.frame.}
\usage{
.sf_from_raster_list(x, time_name)
}
\arguments{
\item{x}{A list of \code{Raster*} objects.}

\item{time_name}{A string giving the name of the time variable.}
}
\value{
An \code{sf} data.frame.
}
\description{
Reform a list of \code{Raster*} objects to an \code{sf} data.frame.
}
