% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_to_input.R
\name{.sf_to_observation}
\alias{.sf_to_observation}
\title{Convert an sf object to a spatio-temporal Observation.}
\usage{
.sf_to_observation(sf_object, Process, formula, n_neighbours, distribution,
  link, silent = T, max_dist = Inf, distance_units = "km", ...)
}
\arguments{
\item{sf_object}{An object of class `sf` containing point geometries and
covariate columns.}

\item{Process}{An object of class `sf` containing point geometries. These
will be used as the parent nodes for the observation locations.}

\item{n_neighbours}{A positive integer giving the (maximum) number of
parents for each node.}

\item{distribution}{A character vector giving the response distribution.
See \code{get_staRVe_distributions("distribution")}.}

\item{link}{A character vector giving the response link function.
See \code{get_staRVe_distributions("link")}.}

\item{silent}{Should intermediate calculations be printed?}

\item{max_dist}{The maximum distance used to search for parents.
Unless this has a units attribute, units are assumed to be the same as
the supplied \code{distance_units}. See \code{\link{construct_dag}}.}

\item{distance_units}{Which units should be used for distances?
See \code{\link{construct_dag}}.}

\item{A}{formula used to describe the model. See \code{prepare_staRVe_input}.}
}
\value{
A list with 4 elements:
  \describe{
    \item{Observation}{An sf object containing point geometries and covariates
       for each observation.}
    \item{data}{A list to pass to TMB::MakeADFun.}
    \item{pars}{A list to pass to TMB::MakeADFun.}
    \item{rand}{A list to pass to TMB::MakeADFun.}
    \item{map}{A list to pass to TMB::MakeADFun.}
 }
}
\description{
Convert an sf object to a spatio-temporal Observation.
}
