% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.sf_to_stars}
\alias{.sf_to_stars}
\title{Convert an sf data.frame to a stars object}
\usage{
.sf_to_stars(x, time_column)
}
\arguments{
\item{x}{An sf object}

\item{time_column}{The name of the column in x giving the time index}
}
\description{
Takes the sf geometry and the time column as the dimensions for a stars object
  with the remaining columns taken as attributes. If any location / time combination
  is not present in x, then the corresponding entry of the stars object will be NA.
Inverse operation of st_as_sf(x,long=T)
}
