% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{.time_from_formula}
\alias{.time_from_formula}
\title{Retrieve a time index from a formula and a data.frame.}
\usage{
.time_from_formula(x, data)
}
\arguments{
\item{x}{A formula object with a time(...,type) function containing one entry,
and a `type' argument which can be one of "ar1", "rw", or "independent".}

\item{data}{A data.frame to retrieve the time index from.}
}
\value{
A vector containing the index, with a `name' attribute
 containing the name of the time index and a `type' attribute containing the
 type of time process (ar1, random walk, or independent).
}
\description{
Retrieve a time index from a formula and a data.frame.
}
