% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_staRVe}
\alias{fit_staRVe}
\title{Fit a TMB model}
\usage{
fit_staRVe(Input, silent = F, runOrderings = F, ...)
}
\arguments{
\item{Input}{A list with three components, the output of \code{\link{prepare_staRVe_input}}.
The first element stores the raw TMB input, the second stores \code{sf} objects
for Observation and Process, and the third stores model settings.}

\item{silent}{Suppress optimization and residual tracing information?}

\item{runOrderings}{Logical value. If true, runs TMB::runSymbolicAnalysis
on the MakeADFun TMB object. Requires a special install of TMB,
see \code{?TMB::runSymbolicAnalysis}.}

\item{...}{Optional parameters to be passed to TMB::MakeADFun, TMB::sdreport,
or nlminb.}
}
\value{
An object of class \code{staRVe} containing the model fit.
}
\description{
Fit a TMB model
}
