% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/dag.R
\name{idx_exchange}
\alias{idx_exchange}
\alias{idxC_to_R}
\alias{idxR_to_C}
\alias{idxC_to_R,dag-method}
\alias{idxR_to_C,dag-method}
\title{Exchange indices between C++ and R}
\usage{
idxC_to_R(x)

idxR_to_C(x)

\S4method{idxC_to_R}{dag}(x)

\S4method{idxR_to_C}{dag}(x)
}
\arguments{
\item{x}{An object with numeric entries}
}
\description{
Converts indices to be used in C++ (where indices start at 0) and R
  (where indices start at 1).
}
\section{Functions}{
\itemize{
\item \code{idxC_to_R}: Add 1 to all indices to convert from C++ to R

\item \code{idxR_to_C}: Subtract 1 from all indices to convert from R to C++

\item \code{idxC_to_R,dag-method}: Add 1 to all vertices in the graph edge list

\item \code{idxR_to_C,dag-method}: Subtract 1 from all vertices in the graph edge list
}}

\keyword{internal}
