% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/long_stars.R
\docType{class}
\name{long_stars-class}
\alias{long_stars-class}
\alias{values,long_stars-method}
\alias{values<-,long_stars-method}
\alias{locations,long_stars-method}
\alias{locations<-,long_stars-method}
\title{An S4 class to hold a multi-dimensional array and a data.frame with a
  many-to-one relationship.}
\usage{
\S4method{values}{long_stars}(x)

\S4method{values}{long_stars}(x) <- value

\S4method{locations}{long_stars}(x)

\S4method{locations}{long_stars}(x) <- value
}
\arguments{
\item{x}{An object}

\item{value}{A replacement value}
}
\description{
An S4 class to hold a multi-dimensional array and a data.frame with a
  many-to-one relationship.
}
\details{
The \code{predictions} slot will have at least 1 dimension, the
  first of which must be named \code{idx} and is identical to the row
  dimension of \code{locations}.
}
\section{Functions}{
\itemize{
\item \code{values,long_stars-method}: Get values in multi-dimensional array

\item \code{values<-,long_stars-method}: Set values in multi-dimensional array

\item \code{locations,long_stars-method}: Get locations

\item \code{locations<-,long_stars-method}: Set locations
}}

\section{Slots}{

\describe{
\item{\code{predictions}}{A stars object.}

\item{\code{locations}}{An sf object with at least a time column and point
geometries.}
}}

\seealso{
Other starve_classes: 
\code{\link{TMB_out-class}},
\code{\link{dag-class}},
\code{\link{observation_parameters-class}},
\code{\link{observations-class}},
\code{\link{parameters-class}},
\code{\link{process-class}},
\code{\link{process_parameters-class}},
\code{\link{settings-class}},
\code{\link{starve-class}},
\code{\link{tracing-class}}
}
\concept{starve_classes}
