% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/observation_parameters.R
\docType{class}
\name{observation_parameters-class}
\alias{observation_parameters-class}
\alias{response_distribution,observation_parameters-method}
\alias{response_distribution<-,observation_parameters-method}
\alias{response_parameters,observation_parameters-method}
\alias{response_parameters<-,observation_parameters-method}
\alias{link_function,observation_parameters-method}
\alias{link_function<-,observation_parameters-method}
\alias{fixed_effects,observation_parameters-method}
\alias{fixed_effects<-,observation_parameters-method}
\title{An S4 class to hold observation parameters for a starve model.}
\usage{
\S4method{response_distribution}{observation_parameters}(x)

\S4method{response_distribution}{observation_parameters}(x) <- value

\S4method{response_parameters}{observation_parameters}(x)

\S4method{response_parameters}{observation_parameters}(x) <- value

\S4method{link_function}{observation_parameters}(x)

\S4method{link_function}{observation_parameters}(x) <- value

\S4method{fixed_effects}{observation_parameters}(x)

\S4method{fixed_effects}{observation_parameters}(x) <- value
}
\arguments{
\item{x}{An object}

\item{value}{A replacement value}
}
\description{
An S4 class to hold observation parameters for a starve model.
}
\section{Functions}{
\itemize{
\item \code{response_distribution,observation_parameters-method}: Get response distribution(s)

\item \code{response_distribution<-,observation_parameters-method}: Get/set response distribution(s). Run
get_starve_distributions("distribution") for valid options. Setting the
the response distribution also overwrites the response parameters and
link function(s).

\item \code{response_parameters,observation_parameters-method}: Get response distribution parameters

\item \code{response_parameters<-,observation_parameters-method}: Set response distribution parameters

\item \code{link_function,observation_parameters-method}: Get link function(s)

\item \code{link_function<-,observation_parameters-method}: set link function(s).  Run
get_starve_distributions("link") for valid options.

\item \code{fixed_effects,observation_parameters-method}: Get fixed effect parameters

\item \code{fixed_effects<-,observation_parameters-method}: Set fixed effect parameters
}}

\section{Slots}{

\describe{
\item{\code{response_distribution}}{The response distribution(s), must be one from
get_starve_distributions("distribution").}

\item{\code{response_parameters}}{A list containing a data.frame of response
distribution parameters for each response variable.}

\item{\code{link_function}}{The link function(s), must be one from
get_starve_distributions("link").}

\item{\code{fixed_effects}}{A list containing a data.frame of fixed effect parameters
for each response variable.}
}}

\seealso{
Other starve_classes: 
\code{\link{TMB_out-class}},
\code{\link{dag-class}},
\code{\link{long_stars-class}},
\code{\link{observations-class}},
\code{\link{parameters-class}},
\code{\link{process-class}},
\code{\link{process_parameters-class}},
\code{\link{settings-class}},
\code{\link{starve-class}},
\code{\link{tracing-class}}
}
\concept{starve_classes}
\keyword{internal}
