% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/staRVe_observations.R
\docType{class}
\name{observations-class}
\alias{observations-class}
\alias{dat,observations-method}
\alias{dat<-,observations-method}
\alias{data_predictions,observations-method}
\alias{data_predictions<-,observations-method}
\alias{parameters,observations-method}
\alias{parameters<-,observations,ANY-method}
\title{An S4 class to hold the observation information for a starve model.}
\usage{
\S4method{dat}{observations}(x)

\S4method{dat}{observations}(x) <- value

\S4method{data_predictions}{observations}(x)

\S4method{data_predictions}{observations}(x) <- value

\S4method{parameters}{observations}(x)

\S4method{parameters}{observations,ANY}(x) <- value
}
\arguments{
\item{x}{An object}

\item{value}{A replacement value}
}
\description{
An S4 class to hold the observation information for a starve model.
}
\section{Functions}{
\itemize{
\item \code{dat,observations-method}: Get data, including response variables,
time indices, locations, covariates, etc.

\item \code{dat<-,observations-method}: Set data. Warning: if you want to add new
rows you must give each additional row a value in the graph_idx column,
or create a new object from scratch.

\item \code{data_predictions,observations-method}: Get data_predictions, a long_stars object
with the data (see \code{dat}) and associated random effect predictions.

\item \code{data_predictions<-,observations-method}: Set data predictions

\item \code{parameters,observations-method}: Get parameters as an
observation_parameters object.

\item \code{parameters<-,observations,ANY-method}: Set parameters using a new
observation_parameters object. Not the recommended way to
modify specific parmaeter values, instead see the package vignette
}}

\section{Slots}{

\describe{
\item{\code{data_predictions}}{A long_stars object for the data.}

\item{\code{transient_graph}}{A dag object describing the dependence of the data on
the process.}

\item{\code{parameters}}{An object of class observation_parameters.}
}}

\seealso{
strv_prepare_observations

Other starve_classes: 
\code{\link{TMB_out-class}},
\code{\link{dag-class}},
\code{\link{long_stars-class}},
\code{\link{observation_parameters-class}},
\code{\link{parameters-class}},
\code{\link{process-class}},
\code{\link{process_parameters-class}},
\code{\link{settings-class}},
\code{\link{starve-class}},
\code{\link{starve_classes}},
\code{\link{tracing-class}}
}
\concept{starve_classes}
\keyword{internal}
