% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag.R
\name{order_by_location}
\alias{order_by_location}
\title{Sort an \code{sf} object with point geometry by location.}
\usage{
order_by_location(x, time = NULL, return = "sort")
}
\arguments{
\item{x}{An \code{sf} object containing point geometries.}

\item{time}{If present, a time index to be the primary (slowest running)
sorting term.}

\item{return}{Should be one of ``sort" or ``order". If ``sort", a sorted copy
of \code{x} is returned. If ``order", a vector of sorted indices is returned.}
}
\value{
Either a copy of \code{x} whose rows as sorted (\code{return='sort'});
 or an integer vector of sorted indices (\code{return='order'}).
}
\description{
The sorting is determined by lexicographic ordering, in order of the columns
 of the geometry object. The locations are put in increasing order according
 to the first column, and then ties are put in order according to the second
 column, and so on. See \code{\link{order}}.
}
