% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staRVe_process.R
\name{prepare_staRVe_process}
\alias{prepare_staRVe_process}
\title{Prepare the process part of a staRVe_model object.}
\usage{
prepare_staRVe_process(
  nodes,
  persistent_graph = NA,
  time = data.frame(time = 0),
  settings = new("staRVe_settings")
)
}
\arguments{
\item{nodes}{An sf object containing the observations and covariates.}

\item{time}{A staRVe_process object.}

\item{settings}{A staRVe_settings object.}
}
\value{
A staRVe_process object.
}
\description{
Prepare the process part of a staRVe_model object.
}
