% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/process.R
\docType{class}
\name{process-class}
\alias{process-class}
\alias{time_effects,process-method}
\alias{time_effects<-,process-method}
\alias{pg_re,process-method}
\alias{pg_re<-,process-method}
\alias{tg_re,process-method}
\alias{tg_re<-,process-method}
\alias{persistent_graph,process-method}
\alias{transient_graph,process-method}
\alias{parameters,process-method}
\alias{parameters<-,process,ANY-method}
\alias{random_effects,process-method}
\title{An S4 class to hold the process information for a starve model.}
\usage{
\S4method{time_effects}{process}(x)

\S4method{time_effects}{process}(x) <- value

\S4method{pg_re}{process}(x)

\S4method{pg_re}{process}(x) <- value

\S4method{tg_re}{process}(x)

\S4method{tg_re}{process}(x) <- value

\S4method{persistent_graph}{process}(x)

\S4method{transient_graph}{process}(x)

\S4method{parameters}{process}(x)

\S4method{parameters}{process,ANY}(x) <- value

\S4method{random_effects}{process}(x)
}
\arguments{
\item{x}{An object}

\item{value}{A replacement value}
}
\description{
An S4 class to hold the process information for a starve model.
}
\section{Functions}{
\itemize{
\item \code{time_effects(process)}: Get temporal random effects

\item \code{time_effects(process) <- value}: Set temporal random effects

\item \code{pg_re(process)}: Get spatio-temporal random effects for persistent
graph

\item \code{pg_re(process) <- value}: Set spatio-temporal random effects for persistent
graph

\item \code{tg_re(process)}: Get spatio-temporal random effects for transient
graph

\item \code{tg_re(process) <- value}: Set spatio-temporal random effects for transient
graph

\item \code{persistent_graph(process)}: Get persistent graph

\item \code{transient_graph(process)}: Get transient graph

\item \code{parameters(process)}: Get parameters as a process_parameters object

\item \code{parameters(x = process) <- value}: Set parameters using a new process_parameters
object. Not the recommended way to modify specific parameter values,
instead see vignette("starve-tour").

\item \code{random_effects(process)}: Get list containing the persistent graph random
effects and the transient graph random effects.

}}
\section{Slots}{

\describe{
\item{\code{time_effects}}{A stars object containing temporal random effects}

\item{\code{random_effects}}{A stars object containing spatio-temporal random
effects.}

\item{\code{persistent_graph}}{A dag object describing the dependence graph of the
process.}

\item{\code{parameters}}{An object of class process_parameters.}
}}

\seealso{
strv_prepare

Other starve_classes: 
\code{\link{TMB_out-class}},
\code{\link{dag-class}},
\code{\link{long_stars-class}},
\code{\link{observation_parameters-class}},
\code{\link{observations-class}},
\code{\link{parameters-class}},
\code{\link{process_parameters-class}},
\code{\link{settings-class}},
\code{\link{starve-class}},
\code{\link{tracing-class}}
}
\concept{starve_classes}
\keyword{internal}
