% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/process_parameters.R
\docType{class}
\name{process_parameters-class}
\alias{process_parameters-class}
\alias{covariance_function,process_parameters-method}
\alias{covariance_function<-,process_parameters-method}
\alias{space_parameters,process_parameters-method}
\alias{space_parameters<-,process_parameters-method}
\alias{time_parameters,process_parameters-method}
\alias{time_parameters<-,process_parameters-method}
\title{An S4 class to hold process parameters for a starve model.}
\usage{
\S4method{covariance_function}{process_parameters}(x)

\S4method{covariance_function}{process_parameters}(x) <- value

\S4method{space_parameters}{process_parameters}(x)

\S4method{space_parameters}{process_parameters}(x) <- value

\S4method{time_parameters}{process_parameters}(x)

\S4method{time_parameters}{process_parameters}(x) <- value
}
\arguments{
\item{x}{An object}

\item{value}{A replacement value}
}
\description{
An S4 class to hold process parameters for a starve model.
}
\section{Functions}{
\itemize{
\item \code{covariance_function(process_parameters)}: Get covariance function(s)

\item \code{covariance_function(process_parameters) <- value}: Set covariance function(s). Run
get_starve_distributions("covariance") for valid covariance functions.
Setting the covariance function also overwrites the spatial parameters.

\item \code{space_parameters(process_parameters)}: Get spatial parameters

\item \code{space_parameters(process_parameters) <- value}: Set spatial parameters

\item \code{time_parameters(process_parameters)}: Get time parameters

\item \code{time_parameters(process_parameters) <- value}: Set time parameters

}}
\section{Slots}{

\describe{
\item{\code{covariance_function}}{The covariance function(s), must be one from
get_starve_distributions("covariance").}

\item{\code{space_parameters}}{A list containing spatial parameters for each response
variable.}

\item{\code{time_parameters}}{A list containing time parameters for each response
variable.}
}}

\seealso{
Other starve_classes: 
\code{\link{TMB_out-class}},
\code{\link{dag-class}},
\code{\link{long_stars-class}},
\code{\link{observation_parameters-class}},
\code{\link{observations-class}},
\code{\link{parameters-class}},
\code{\link{process-class}},
\code{\link{settings-class}},
\code{\link{starve-class}},
\code{\link{tracing-class}}
}
\concept{starve_classes}
\keyword{internal}
