% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag.R
\name{Rdag_to_Cdag}
\alias{Rdag_to_Cdag}
\title{Reduce all the parent indices of a directed acyclic graph by one for use in cpp.}
\usage{
Rdag_to_Cdag(dag)
}
\arguments{
\item{dag}{A directed acyclic graph. See \code{\link{construct_dag}}.}
}
\value{
The same directed acyclic graph as passed in with all parent indices
 reduced by one.
}
\description{
Reduce all the parent indices of a directed acyclic graph by one for use in cpp.
}
