% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/settings.R
\docType{class}
\name{settings-class}
\alias{settings-class}
\alias{formula,settings-method}
\alias{formula<-,settings-method}
\alias{n_neighbours,settings-method}
\alias{distance_units,settings-method}
\alias{distance_units<-,settings-method}
\alias{max_distance,settings-method}
\alias{max_distance<-,settings-method}
\title{An S4 class to hold settings for a starve model}
\usage{
\S4method{formula}{settings}(x)

\S4method{formula}{settings}(x) <- value

\S4method{n_neighbours}{settings}(x)

\S4method{distance_units}{settings}(x)

\S4method{distance_units}{settings}(x) <- value

\S4method{max_distance}{settings}(x)

\S4method{max_distance}{settings}(x) <- value
}
\arguments{
\item{x}{An object}

\item{value}{A replacement value}
}
\description{
An S4 class to hold settings for a starve model
}
\section{Functions}{
\itemize{
\item \code{formula(settings)}: Get model formula

\item \code{formula(settings) <- value}: Set model formula

\item \code{n_neighbours(settings)}: Get maximum number of neighbours used in graphs

\item \code{distance_units(settings)}: Get distance units used for the model.

\item \code{distance_units(settings) <- value}: Set distance units. Automatically converts
max_distance setting to new units.

\item \code{max_distance(settings)}: Get maximum distance for edges in the graph

\item \code{max_distance(settings) <- value}: Set maximum distance for edges in the graph

}}
\section{Slots}{

\describe{
\item{\code{formula}}{A formula}

\item{\code{n_neighbours}}{The number of parents for each node in the graph.}

\item{\code{distance_units}}{The units used for distance calculation.  Must be
compatible with units::set_units.}

\item{\code{max_distance}}{The maximum allowable distance for edges in the transient
graph.}
}}

\seealso{
Other starve_classes: 
\code{\link{TMB_out-class}},
\code{\link{dag-class}},
\code{\link{long_stars-class}},
\code{\link{observation_parameters-class}},
\code{\link{observations-class}},
\code{\link{parameters-class}},
\code{\link{process-class}},
\code{\link{process_parameters-class}},
\code{\link{starve-class}},
\code{\link{tracing-class}}
}
\concept{starve_classes}
\keyword{internal}
