% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\name{staRVe-access}
\alias{staRVe-access}
\alias{convergence}
\alias{covariance_function}
\alias{covariance_function<-}
\alias{dat}
\alias{dat<-}
\alias{data_predictions}
\alias{data_predictions<-}
\alias{distance_units}
\alias{distance_units<-}
\alias{distances}
\alias{edges}
\alias{extras}
\alias{extras<-}
\alias{fixed_effects}
\alias{fixed_effects<-}
\alias{formula}
\alias{formula<-}
\alias{graph}
\alias{hess_time}
\alias{link_function}
\alias{link_function<-}
\alias{locations}
\alias{locations<-}
\alias{max_distance}
\alias{n_neighbours}
\alias{obj}
\alias{obs_dag_method}
\alias{opt}
\alias{opt_time}
\alias{p_far_neighbours}
\alias{parameter_covariance}
\alias{parameter_hessian}
\alias{parameters}
\alias{parameters<-}
\alias{pg_re}
\alias{pg_re<-}
\alias{predictions}
\alias{predictions<-}
\alias{random_effects}
\alias{random_effects<-}
\alias{response_distribution}
\alias{response_distribution<-}
\alias{response_parameters}
\alias{response_parameters<-}
\alias{sdr}
\alias{sdr_time}
\alias{settings}
\alias{spatial_parameters}
\alias{spatial_parameters<-}
\alias{tg_re}
\alias{tg_re<-}
\alias{time_effects}
\alias{time_effects<-}
\alias{time_parameters}
\alias{time_parameters<-}
\alias{timing}
\alias{TMB_out}
\alias{tracing}
\alias{values}
\alias{values<-}
\title{Access methods for the staRVe package}
\usage{
convergence(x)

covariance_function(x)

covariance_function(x) <- value

dat(x)

dat(x) <- value

data_predictions(x, ...)

data_predictions(x) <- value

distance_units(x)

distance_units(x) <- value

distances(x)

edges(x)

extras(x)

extras(x) <- value

fixed_effects(x)

fixed_effects(x) <- value

formula(x, ...)

formula(x) <- value

graph(x)

hess_time(x)

link_function(x)

link_function(x) <- value

locations(x)

locations(x) <- value

max_distance(x)

n_neighbours(x)

obj(x)

obs_dag_method(x)

opt(x)

opt_time(x)

p_far_neighbours(x)

parameter_covariance(x)

parameter_hessian(x)

parameters(x)

parameters(x) <- value

pg_re(x)

pg_re(x) <- value

predictions(x)

predictions(x) <- value

random_effects(x)

random_effects(x) <- value

response_distribution(x)

response_distribution(x) <- value

response_parameters(x)

response_parameters(x) <- value

sdr(x)

sdr_time(x)

settings(x)

spatial_parameters(x)

spatial_parameters(x) <- value

tg_re(x)

tg_re(x) <- value

time_effects(x)

time_effects(x) <- value

time_parameters(x)

time_parameters(x) <- value

timing(x)

TMB_out(x)

tracing(x)

values(x)

values(x) <- value
}
\arguments{
\item{x}{An object}

\item{value}{A replacement value}

\item{...}{Not used}
}
\description{
Access methods for the staRVe package
}
\section{Functions}{
\itemize{
\item \code{convergence}: Get convergence message

\item \code{covariance_function}: Get covariance function

\item \code{covariance_function<-}: Set covariance function

\item \code{dat}: Get data

\item \code{dat<-}: Set data

\item \code{data_predictions}: Get data predictions

\item \code{data_predictions<-}: Set data

\item \code{distance_units}: Get distance units

\item \code{distance_units<-}: Set distance units

\item \code{distances}: Get distances

\item \code{edges}: Get edges

\item \code{extras}: Get extra settings

\item \code{extras<-}: Set extra settings

\item \code{fixed_effects}: Get fixed effects

\item \code{fixed_effects<-}: Set fixed effects

\item \code{formula}: Get formula

\item \code{formula<-}: Set formula

\item \code{graph}: Get graph

\item \code{hess_time}: Get hessian time

\item \code{link_function}: Get link function

\item \code{link_function<-}: Set link function

\item \code{locations}: Get locations

\item \code{locations<-}: Get locations

\item \code{max_distance}: Get max distance

\item \code{n_neighbours}: Get n neighbours

\item \code{obj}: Get obj

\item \code{obs_dag_method}: Get obs_dag_method

\item \code{opt}: Get opt

\item \code{opt_time}: Get opt time

\item \code{p_far_neighbours}: Get p far neighbours

\item \code{parameter_covariance}: Get parameter covariance

\item \code{parameter_hessian}: Get parameter hessian

\item \code{parameters}: Get parameters

\item \code{parameters<-}: Set parameters

\item \code{pg_re}: Get persistent graph random effects

\item \code{pg_re<-}: Set persistent graph random effects

\item \code{predictions}: Get predictions

\item \code{predictions<-}: Set predictions

\item \code{random_effects}: Get random effects

\item \code{random_effects<-}: Set random effects

\item \code{response_distribution}: Get response distribution

\item \code{response_distribution<-}: Set response distribution

\item \code{response_parameters}: Get response parameters

\item \code{response_parameters<-}: Set response parameters

\item \code{sdr}: Get sdr

\item \code{sdr_time}: Get sdr time

\item \code{settings}: Get settings

\item \code{spatial_parameters}: Get spatial parameters

\item \code{spatial_parameters<-}: Set spatial parameters

\item \code{tg_re}: Get transient graph random effects

\item \code{tg_re<-}: Set transient graph random effects

\item \code{time_effects}: Get time effects

\item \code{time_effects<-}: Set time effects

\item \code{time_parameters}: Get time parameters

\item \code{time_parameters<-}: Set time parameters

\item \code{timing}: Get timing

\item \code{TMB_out}: Get TMB_out

\item \code{tracing}: Get tracing

\item \code{values}: Get values

\item \code{values<-}: Set values
}}

