% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{staRVe-class}
\alias{staRVe-class}
\title{An S4 class to hold results from an analysis of RV data.}
\description{
An S4 class to hold results from an analysis of RV data.
}
\section{Slots}{

\describe{
\item{\code{TMB_out}}{An object of class \code{TMB_out} holding the raw TMB output of the model.}

\item{\code{Observation}}{An \code{sf} object containing data.frame with columns \code{y}, \code{response}, \code{response_se}, and \code{residual}, and an \code{sfc} object containing location data.}

\item{\code{Process}}{An \code{sf} object containing data.frame with columns \code{w} and \code{w_se}, and an \code{sfc} object containing location data.}

\item{\code{parameters}}{A data.frame containing the columns \code{par}, and \code{par_se}.}

\item{\code{convergence}}{A character vector giving the convergence message.}
}}

\concept{staRVe-classe}
