% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/TMB_out.R, R/dag.R,
%   R/long_stars.R, R/staRVe_tracing.R, R/staRVe_settings.R,
%   R/staRVe_observation_parameters.R, R/staRVe_process_parameters.R,
%   R/staRVe_parameters.R, R/staRVe_process.R, R/staRVe_observations.R,
%   R/staRVe_model.R
\name{starve-construct}
\alias{starve-construct}
\alias{initialize,TMB_out-method}
\alias{initialize,dag-method}
\alias{initialize,long_stars-method}
\alias{initialize,tracing-method}
\alias{initialize,settings-method}
\alias{initialize,observation_parameters-method}
\alias{initialize,process_parameters-method}
\alias{initialize,parameters-method}
\alias{initialize,process-method}
\alias{initialize,observations-method}
\alias{initialize,starve-method}
\title{Constructors for the starve package}
\usage{
\S4method{initialize}{TMB_out}(
  .Object,
  obj = list(),
  opt = list(convergence = 0, message = "Model not yet fitted."),
  sdr = structure(numeric(0), class = "sdreport")
)

\S4method{initialize}{dag}(.Object, edges = list(), distances = list(), distance_units = "km")

\S4method{initialize}{long_stars}(
  .Object,
  locations = sf::st_sf(data.frame(time = numeric(0)), geometry =
    sf::st_sfc(sf::st_point())[-1]),
  var_names = "y"
)

\S4method{initialize}{tracing}(
  .Object,
  opt_time = proc.time() - proc.time(),
  hess_time = proc.time() - proc.time(),
  sdr_time = proc.time() - proc.time(),
  parameter_hessian = matrix(numeric(0), nrow = 0, ncol = 0),
  parameter_covariance = matrix(numeric(0), nrow = 0, ncol = 0)
)

\S4method{initialize}{settings}(
  .Object,
  formula = new("formula"),
  n_neighbours = 10,
  distance_units = "km",
  max_distance = Inf
)

\S4method{initialize}{observation_parameters}(
  .Object,
  response_distribution = "gaussian",
  fixed_effects = list(data.frame(par = numeric(0), fixed = numeric(0)))
)

\S4method{initialize}{process_parameters}(
  .Object,
  covariance_function = "exponential",
  time_parameters = list(data.frame(par = c(0, 0, 0), se = c(0, 0, 0), fixed = c(FALSE,
    FALSE, FALSE), row.names = c("mu", "ar1", "sd")))
)

\S4method{initialize}{parameters}(
  .Object,
  process_parameters = new("process_parameters"),
  observation_parameters = new("observation_parameters")
)

\S4method{initialize}{process}(
  .Object,
  time_effects = stars::st_as_stars(list(w = array(0, dim = c(1, 1)), se = array(0, dim
    = c(1, 1))), dimensions = stars::st_dimensions(time = 0, variable = "y")),
  pg_re = stars::st_as_stars(list(w = array(0, dim = c(1, 1, 1)), se = array(0, dim =
    c(1, 1, 1))), dimensions = stars::st_dimensions(geom = sf::st_sfc(sf::st_point(c(0,
    0))), time = 0, variable = "y")),
  tg_re = new("long_stars"),
  persistent_graph = new("dag"),
  transient_graph = new("dag"),
  parameters = new("process_parameters")
)

\S4method{initialize}{observations}(
  .Object,
  data_predictions = new("long_stars"),
  parameters = new("observation_parameters")
)

\S4method{initialize}{starve}(
  .Object,
  process = new("process"),
  observations = new("observations"),
  tracing = new("tracing"),
  TMB_out = new("TMB_out"),
  settings = new("settings")
)
}
\arguments{
\item{.Object}{The object to create}

\item{obj}{Output of TMB::MakeADFun}

\item{opt}{Output of nlminb}

\item{sdr}{Output of TMB::sdreport}

\item{edges}{A list of edges}

\item{distances}{A list of distances}

\item{distance_units}{Which distance units to use}

\item{locations}{An sf object with a time column and point geometries}

\item{var_names}{A character vector}

\item{opt_time}{A proc_time object}

\item{hess_time}{A proc_time object}

\item{sdr_time}{A proc_time object}

\item{parameter_hessian}{A matrix}

\item{parameter_covariance}{A matrix}

\item{formula}{A formula}

\item{n_neighbours}{An integer}

\item{response_distribution}{Which response distribution to use}

\item{fixed_effects}{A data.frame}

\item{covariance_function}{Which covariance function to use}

\item{time_parameters}{A data.frame}

\item{process_parameters}{A process_parameters object}

\item{observation_parameters}{A observation_parameters object}

\item{time_effects}{A stars object}

\item{pg_re}{A stars object, random effects for the persistent graph}

\item{tg_re}{A stars object, random effects for the transient graph}

\item{persistent_graph}{A dag object}

\item{transient_graph}{A dag object}

\item{parameters}{An observation_parameters object}

\item{data_predictions}{A long_stars object}

\item{process}{A process object}

\item{observations}{A observations object}

\item{tracing}{A tracing object}

\item{TMB_out}{A TMB_out object}

\item{settings}{A settings object}
}
\description{
Constructors for the starve package
}
\keyword{internal}
