% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_staRVe.R
\name{staRVe}
\alias{staRVe}
\alias{initialize,staRVe-method}
\title{Create a \code{staRVe} object.}
\usage{
staRVe(x, ...)

\S4method{initialize}{staRVe}(
  .Object,
  TMB_out = new("TMB_out"),
  observation = sf::st_sf(data.frame(y = numeric(1), response = numeric(1), response_se
    = numeric(1), residual = numeric(1)), geometry = sf::st_sfc(sf::st_point())),
  process = sf::st_sf(data.frame(w = numeric(1), w_se = numeric(1)), geometry =
    sf::st_sfc(sf::st_point())),
  parameters = data.frame(parameter = character(0), estimate = numeric(0), se =
    numeric(0)),
  convergence = character(0),
  settings = list(n_neighbours = numeric(0), distance_units = structure(numeric(0),
    class = "units"))
)
}
\arguments{
\item{x}{An object to coerce to class \code{staRVe}.}
}
\description{
Create a \code{staRVe} object.
}
\details{
The \code{initialize} function is not meant to be used by the user,
 use \code{staRVe} instead.
}
