% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\name{starve_access}
\alias{starve_access}
\alias{convergence}
\alias{covariance_function}
\alias{covariance_function<-}
\alias{dat}
\alias{dat<-}
\alias{data_predictions}
\alias{data_predictions<-}
\alias{distance_units}
\alias{distance_units<-}
\alias{distances}
\alias{edges}
\alias{fixed_effects}
\alias{fixed_effects<-}
\alias{formula}
\alias{formula<-}
\alias{graph}
\alias{hess_time}
\alias{link_function}
\alias{link_function<-}
\alias{locations}
\alias{locations<-}
\alias{max_distance}
\alias{max_distance<-}
\alias{n_neighbours}
\alias{opt_time}
\alias{parameter_covariance}
\alias{parameter_hessian}
\alias{parameters}
\alias{parameters<-}
\alias{persistent_graph}
\alias{pg_re}
\alias{pg_re<-}
\alias{random_effects}
\alias{response_distribution}
\alias{response_distribution<-}
\alias{response_parameters}
\alias{response_parameters<-}
\alias{sdr_time}
\alias{settings}
\alias{space_parameters}
\alias{space_parameters<-}
\alias{tg_re}
\alias{tg_re<-}
\alias{time_effects}
\alias{time_effects<-}
\alias{time_parameters}
\alias{time_parameters<-}
\alias{timing}
\alias{tracing}
\alias{transient_graph}
\alias{values}
\alias{values<-}
\title{A list of accessor functions for the starve package}
\usage{
convergence(x)

covariance_function(x)

covariance_function(x) <- value

dat(x)

dat(x) <- value

data_predictions(x, ...)

data_predictions(x) <- value

distance_units(x)

distance_units(x) <- value

distances(x)

edges(x)

fixed_effects(x)

fixed_effects(x) <- value

formula(x, ...)

formula(x) <- value

graph(x)

hess_time(x)

link_function(x)

link_function(x) <- value

locations(x)

locations(x) <- value

max_distance(x)

max_distance(x) <- value

n_neighbours(x)

opt_time(x)

parameter_covariance(x)

parameter_hessian(x)

parameters(x)

parameters(x) <- value

persistent_graph(x)

pg_re(x)

pg_re(x) <- value

random_effects(x)

response_distribution(x)

response_distribution(x) <- value

response_parameters(x)

response_parameters(x) <- value

sdr_time(x)

settings(x)

space_parameters(x)

space_parameters(x) <- value

tg_re(x)

tg_re(x) <- value

time_effects(x)

time_effects(x) <- value

time_parameters(x)

time_parameters(x) <- value

timing(x)

tracing(x)

transient_graph(x)

values(x)

values(x) <- value
}
\arguments{
\item{x}{An object}

\item{value}{A replacement value}
}
\description{
These accessor functions are used to extract different components from the
  objects used in the starve package. This help page is meant only to catalogue
  all the accessor functions available, and as such will be of limited benefit
  to package users. Not every accessor function is available for
  every class implemented in the package. To see which functions are available
  for a class, and a description of the different components of that class
  that these functions return, please see the documentation for that class.
  The main classes implemented in the starve package are \linkS4class{dag},
  \linkS4class{long_stars}, \linkS4class{tracing}, and \linkS4class{starve}.
}
\section{Functions}{
\itemize{
\item \code{convergence}: Get convergence message

\item \code{covariance_function}: Get covariance function

\item \code{covariance_function<-}: Set covariance function

\item \code{dat}: Get data

\item \code{dat<-}: Set data

\item \code{data_predictions}: Get data predictions

\item \code{data_predictions<-}: Set data

\item \code{distance_units}: Get distance units

\item \code{distance_units<-}: Set distance units

\item \code{distances}: Get distances

\item \code{edges}: Get edges

\item \code{fixed_effects}: Get fixed effects

\item \code{fixed_effects<-}: Set fixed effects

\item \code{formula}: Get model formula

\item \code{formula<-}: Set model formula

\item \code{graph}: Get graph

\item \code{hess_time}: Get hessian computation time

\item \code{link_function}: Get link function

\item \code{link_function<-}: Set link function

\item \code{locations}: Get locations

\item \code{locations<-}: Get locations

\item \code{max_distance}: Get max distance

\item \code{max_distance<-}: Set max distance

\item \code{n_neighbours}: Get n_neighbours setting

\item \code{opt_time}: Get optimization time

\item \code{parameter_covariance}: Get parameter estimator covariance matrix

\item \code{parameter_hessian}: Get parameter estimator hessian matrix

\item \code{parameters}: Get parameters

\item \code{parameters<-}: Set parameters

\item \code{persistent_graph}: Get persistent graph

\item \code{pg_re}: Get persistent graph random effects

\item \code{pg_re<-}: Set persistent graph random effects

\item \code{random_effects}: Get random effects

\item \code{response_distribution}: Get response distribution

\item \code{response_distribution<-}: Set response distribution

\item \code{response_parameters}: Get response distribution parameters

\item \code{response_parameters<-}: Set response distribution parameters

\item \code{sdr_time}: Get standard error computation time

\item \code{settings}: Get settings

\item \code{space_parameters}: Get spatial parameters

\item \code{space_parameters<-}: Set spatial parameters

\item \code{tg_re}: Get transient graph random effects

\item \code{tg_re<-}: Set transient graph random effects

\item \code{time_effects}: Get time effects

\item \code{time_effects<-}: Set time effects

\item \code{time_parameters}: Get time parameters

\item \code{time_parameters<-}: Set time parameters

\item \code{timing}: Get model computation times

\item \code{tracing}: Get tracing information

\item \code{transient_graph}: Get transient graph

\item \code{values}: Get values

\item \code{values<-}: Set values
}}

\seealso{
\linkS4class{dag}, \linkS4class{long_stars}, \linkS4class{tracing}, \linkS4class{starve}
}
\keyword{internal}
