% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/starve.R, R/show.R
\docType{class}
\name{starve-class}
\alias{starve-class}
\alias{settings,starve-method}
\alias{tracing,starve-method}
\alias{time_effects,starve-method}
\alias{time_effects<-,starve-method}
\alias{pg_re,starve-method}
\alias{pg_re<-,starve-method}
\alias{tg_re,starve-method}
\alias{tg_re<-,starve-method}
\alias{random_effects,starve-method}
\alias{persistent_graph,starve-method}
\alias{transient_graph,starve-method}
\alias{graph,starve-method}
\alias{covariance_function,starve-method}
\alias{covariance_function<-,starve-method}
\alias{space_parameters,starve-method}
\alias{space_parameters<-,starve-method}
\alias{time_parameters,starve-method}
\alias{time_parameters<-,starve-method}
\alias{dat,starve-method}
\alias{dat<-,starve-method}
\alias{data_predictions,starve-method}
\alias{data_predictions<-,starve-method}
\alias{response_distribution,starve-method}
\alias{response_distribution<-,starve-method}
\alias{response_parameters,starve-method}
\alias{response_parameters<-,starve-method}
\alias{link_function,starve-method}
\alias{link_function<-,starve-method}
\alias{fixed_effects,starve-method}
\alias{fixed_effects<-,starve-method}
\alias{distance_units,starve-method}
\alias{distance_units<-,starve-method}
\alias{formula,starve-method}
\alias{formula<-,starve-method}
\alias{parameters,starve-method}
\alias{parameters<-,starve,parameters-method}
\alias{convergence,starve-method}
\alias{timing,starve-method}
\alias{show,starve-method}
\title{An S4 class describing a starve model.}
\usage{
\S4method{settings}{starve}(x)

\S4method{tracing}{starve}(x)

\S4method{time_effects}{starve}(x)

\S4method{time_effects}{starve}(x) <- value

\S4method{pg_re}{starve}(x)

\S4method{pg_re}{starve}(x) <- value

\S4method{tg_re}{starve}(x)

\S4method{tg_re}{starve}(x) <- value

\S4method{random_effects}{starve}(x)

\S4method{persistent_graph}{starve}(x)

\S4method{transient_graph}{starve}(x)

\S4method{graph}{starve}(x)

\S4method{covariance_function}{starve}(x)

\S4method{covariance_function}{starve}(x) <- value

\S4method{space_parameters}{starve}(x)

\S4method{space_parameters}{starve}(x) <- value

\S4method{time_parameters}{starve}(x)

\S4method{time_parameters}{starve}(x) <- value

\S4method{dat}{starve}(x)

\S4method{dat}{starve}(x) <- value

\S4method{data_predictions}{starve}(x)

\S4method{data_predictions}{starve}(x) <- value

\S4method{response_distribution}{starve}(x)

\S4method{response_distribution}{starve}(x) <- value

\S4method{response_parameters}{starve}(x)

\S4method{response_parameters}{starve}(x) <- value

\S4method{link_function}{starve}(x)

\S4method{link_function}{starve}(x) <- value

\S4method{fixed_effects}{starve}(x)

\S4method{fixed_effects}{starve}(x) <- value

\S4method{distance_units}{starve}(x)

\S4method{distance_units}{starve}(x) <- value

\S4method{formula}{starve}(x)

\S4method{formula}{starve}(x) <- value

\S4method{parameters}{starve}(x)

\S4method{parameters}{starve,parameters}(x) <- value

\S4method{convergence}{starve}(x)

\S4method{timing}{starve}(x)

\S4method{show}{starve}(object)
}
\arguments{
\item{x}{An object}

\item{value}{A replacement value}

\item{object}{An object}
}
\description{
An S4 class describing a starve model.
}
\section{Functions}{
\itemize{
\item \code{settings(starve)}: Get model settings

\item \code{tracing(starve)}: Get tracing information, see \link{tracing}.

\item \code{time_effects(starve)}: Get temporal random effects

\item \code{time_effects(starve) <- value}: Set temporal random effects

\item \code{pg_re(starve)}: Get persistent graph random effects

\item \code{pg_re(starve) <- value}: Set spatio-temporal random effects

\item \code{tg_re(starve)}: Get transient graph random effects

\item \code{tg_re(starve) <- value}: Set transient graph random effects

\item \code{random_effects(starve)}: Get list containing the persistent graph random
effects and the transient graph random effects.

\item \code{persistent_graph(starve)}: Get persistent graph

\item \code{transient_graph(starve)}: Get transient graph

\item \code{graph(starve)}: Get a list containing the persistent and transient graphs

\item \code{covariance_function(starve)}: Get covariance function

\item \code{covariance_function(starve) <- value}: Set covariance function(s). Run
get_starve_distributions("covariance") for valid covariance functions.
Setting the covariance function also overwrites the spatial parameters.

\item \code{space_parameters(starve)}: Get spatial parameters

\item \code{space_parameters(starve) <- value}: Set spatial parameters

\item \code{time_parameters(starve)}: Get time parameters

\item \code{time_parameters(starve) <- value}: Set time parameters

\item \code{dat(starve)}: Get data including response variables, time indices,
locations, covariates, etc.

\item \code{dat(starve) <- value}: Set data. Warning: if you want to add new
rows you must give each additional row a value in the graph_idx column,
or create a new object from scratch.

\item \code{data_predictions(starve)}: Get data predictions, a long_stars object
with the data (see \code{dat}) and associated random effect
predictions.

\item \code{data_predictions(starve) <- value}: Set data predictions

\item \code{response_distribution(starve)}: Get the response distribution(s)

\item \code{response_distribution(starve) <- value}: Set the response distribution(s). Run
get_starve_distributions("distribution") for valid options.
Setting the response distribution also overwrites the response
parameters and link function(s).

\item \code{response_parameters(starve)}: Get response distribution parameters

\item \code{response_parameters(starve) <- value}: Set response distribution parameters

\item \code{link_function(starve)}: Get link function(s)

\item \code{link_function(starve) <- value}: Set link function(s). Run
get_starve_distributions("link") for valid link functions.

\item \code{fixed_effects(starve)}: Get fixed effect parameters

\item \code{fixed_effects(starve) <- value}: Set fixed effect parameters

\item \code{distance_units(starve)}: Get distance units used for the model

\item \code{distance_units(starve) <- value}: Set distance units used for the model

\item \code{formula(starve)}: Get formula used for model

\item \code{formula(starve) <- value}: Set formula used for the model

\item \code{parameters(starve)}: Get all model parameters as a parameters object

\item \code{parameters(x = starve) <- value}: Set all model parameters with a new
parameters object

\item \code{convergence(starve)}: Get convergence message from optimizer

\item \code{timing(starve)}: Get all timing information, see \link{tracing}

\item \code{show(starve)}: Print method

}}
\section{Slots}{

\describe{
\item{\code{process}}{A process object.}

\item{\code{observations}}{An observations object.}

\item{\code{settings}}{A settings object.}
}}

\seealso{
strv_prepare

Other starve_classes: 
\code{\link{TMB_out-class}},
\code{\link{dag-class}},
\code{\link{long_stars-class}},
\code{\link{observation_parameters-class}},
\code{\link{observations-class}},
\code{\link{parameters-class}},
\code{\link{process-class}},
\code{\link{process_parameters-class}},
\code{\link{settings-class}},
\code{\link{tracing-class}}
}
\concept{starve_classes}
