% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{staRVe_fit-class}
\alias{staRVe_fit-class}
\title{An S4 class to hold an optimized staRVe model}
\description{
An S4 class to hold an optimized staRVe model
}
\section{Slots}{

\describe{
\item{\code{tracing}}{A staRVe_tracing object.}

\item{\code{TMB_out}}{A TMB_out object.}
}}

