% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staRVe_fit.R
\name{staRVe_fit,staRVe_model-method}
\alias{staRVe_fit,staRVe_model-method}
\title{Fit a \code{staRVe_model} object.}
\usage{
\S4method{staRVe_fit}{staRVe_model}(x, silent = F, ...)
}
\description{
Fit a \code{staRVe_model} object.
}
