% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/staRVe_fit.R
\name{staRVe_fit}
\alias{staRVe_fit}
\alias{initialize,staRVe_fit-method}
\title{Create a \code{staRVe_fit} object}
\usage{
staRVe_fit(x, ...)

\S4method{initialize}{staRVe_fit}(
  .Object,
  staRVe_model = new("staRVe_model"),
  tracing = new("staRVe_tracing"),
  TMB_out = new("TMB_out")
)
}
\description{
Create a \code{staRVe_fit} object
}
\details{
The \code{initialize} function is not mean to be used by the user,
  use \code{staRVe_fit} instead.
}
