% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/show.R, R/staRVe_model_fit.R
\docType{class}
\name{starve_fit-class}
\alias{starve_fit-class}
\alias{show,starve_fit-method}
\alias{tracing,starve_fit-method}
\alias{tracing<-,starve_fit-method}
\alias{TMB_out,starve_fit-method}
\alias{TMB_out<-,starve_fit-method}
\alias{convergence,starve_fit-method}
\alias{timing,starve_fit-method}
\title{An extension of the starve class to hold optimization information}
\usage{
\S4method{show}{starve_fit}(object)

\S4method{tracing}{starve_fit}(x)

\S4method{tracing}{starve_fit}(x) <- value

\S4method{TMB_out}{starve_fit}(x)

\S4method{TMB_out}{starve_fit}(x) <- value

\S4method{convergence}{starve_fit}(x)

\S4method{timing}{starve_fit}(x)
}
\arguments{
\item{object}{An object}

\item{x}{An object}

\item{value}{A replacement value}
}
\description{
An extension of the starve class to hold optimization information
}
\section{Functions}{
\itemize{
\item \code{show,starve_fit-method}: Print method

\item \code{tracing,starve_fit-method}: Get tracing information, see \link{tracing_class}.

\item \code{tracing<-,starve_fit-method}: Set tracing information (for internal use only)

\item \code{TMB_out,starve_fit-method}: Get TMB objects (for internal use only)

\item \code{TMB_out<-,starve_fit-method}: Set TMB objects (for internal use only)

\item \code{convergence,starve_fit-method}: Get convergence message from optimizer

\item \code{timing,starve_fit-method}: Get all timing information, see \link{tracing}
}}

\section{Slots}{

\describe{
\item{\code{tracing}}{A tracing object}

\item{\code{TMB_out}}{A TMB_out object (for internal use)}
}}

\seealso{
starve_class, tracing_class, TMB_out

Other starve_classes: 
\code{\link{TMB_out-class}},
\code{\link{dag-class}},
\code{\link{long_stars-class}},
\code{\link{observation_parameters-class}},
\code{\link{observations-class}},
\code{\link{parameters-class}},
\code{\link{process-class}},
\code{\link{process_parameters-class}},
\code{\link{settings-class}},
\code{\link{starve-class}},
\code{\link{starve_classes}},
\code{\link{tracing-class}}
}
\concept{starve_classes}
