% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{staRVe_generics}
\alias{staRVe_generics}
\alias{convergence}
\alias{covariance_function}
\alias{covariance_function<-}
\alias{dag}
\alias{data}
\alias{data<-}
\alias{distance_units}
\alias{distance_units<-}
\alias{distances}
\alias{distances<-}
\alias{edges}
\alias{edges<-}
\alias{fixed_effects}
\alias{fixed_effects<-}
\alias{formula}
\alias{formula<-}
\alias{graph}
\alias{hess_time}
\alias{hess_time<-}
\alias{idxC_to_R}
\alias{idxR_to_C}
\alias{link_function}
\alias{link_function<-}
\alias{max_distance}
\alias{max_distance<-}
\alias{n_neighbours}
\alias{n_neighbours<-}
\alias{obj}
\alias{obj<-}
\alias{observations}
\alias{observations<-}
\alias{opt}
\alias{opt<-}
\alias{opt_time}
\alias{opt_time<-}
\alias{p_far_neighbours}
\alias{p_far_neighbours<-}
\alias{parameter_covariance}
\alias{parameter_covariance<-}
\alias{parameter_hessian}
\alias{parameter_hessian<-}
\alias{parameters}
\alias{parameters<-}
\alias{persistent_graph}
\alias{persistent_graph<-}
\alias{process}
\alias{process<-}
\alias{random_effects}
\alias{random_effects<-}
\alias{report}
\alias{response_distribution}
\alias{response_distribution<-}
\alias{response_parameters}
\alias{response_parameters<-}
\alias{sdr}
\alias{sdr<-}
\alias{sdr_time}
\alias{sdr_time<-}
\alias{sdreport}
\alias{settings}
\alias{settings<-}
\alias{spatial_parameters}
\alias{spatial_parameters<-}
\alias{staRVe_fit}
\alias{staRVe_model}
\alias{staRVe_observation_parameters}
\alias{staRVe_observations}
\alias{staRVe_predict}
\alias{staRVe_process}
\alias{staRVe_process_parameters}
\alias{staRVe_settings}
\alias{staRVe_tracing}
\alias{symbolicAnalysis}
\alias{symbolicAnalysis<-}
\alias{time_parameters}
\alias{time_parameters<-}
\alias{timing}
\alias{TMB_in}
\alias{TMB_out}
\alias{TMB_out<-}
\alias{tracing}
\alias{tracing<-}
\alias{transient_graph}
\alias{transient_graph<-}
\alias{update_staRVe_model}
\title{A list of generics for the staRVe package.}
\usage{
convergence(x)

covariance_function(x)

covariance_function(x) <- value

dag(x, ...)

data(x)

data(x) <- value

distance_units(x)

distance_units(x) <- value

distances(x)

distances(x) <- value

edges(x)

edges(x) <- value

fixed_effects(x)

fixed_effects(x) <- value

formula(x, ...)

formula(x) <- value

graph(x)

hess_time(x)

hess_time(x) <- value

idxC_to_R(x)

idxR_to_C(x)

link_function(x)

link_function(x) <- value

max_distance(x)

max_distance(x) <- value

n_neighbours(x)

n_neighbours(x) <- value

obj(x)

obj(x) <- value

observations(x)

observations(x) <- value

opt(x)

opt(x) <- value

opt_time(x)

opt_time(x) <- value

p_far_neighbours(x)

p_far_neighbours(x) <- value

parameter_covariance(x)

parameter_covariance(x) <- value

parameter_hessian(x)

parameter_hessian(x) <- value

parameters(x)

parameters(x) <- value

persistent_graph(x)

persistent_graph(x) <- value

process(x)

process(x) <- value

random_effects(x)

random_effects(x) <- value

report(x)

response_distribution(x)

response_distribution(x) <- value

response_parameters(x)

response_parameters(x) <- value

sdr(x)

sdr(x) <- value

sdr_time(x)

sdr_time(x) <- value

sdreport(x)

settings(x)

settings(x) <- value

spatial_parameters(x)

spatial_parameters(x) <- value

staRVe_fit(x, ...)

staRVe_model(x, ...)

staRVe_observation_parameters(x, ...)

staRVe_observations(x, ...)

staRVe_predict(x, locations, ...)

staRVe_process(x, ...)

staRVe_process_parameters(x, ...)

staRVe_settings(x, ...)

staRVe_tracing(x, ...)

symbolicAnalysis(x)

symbolicAnalysis(x) <- value

time_parameters(x)

time_parameters(x) <- value

timing(x)

TMB_in(x)

TMB_out(x)

TMB_out(x) <- value

tracing(x)

tracing(x) <- value

transient_graph(x)

transient_graph(x) <- value

update_staRVe_model(x, y)
}
\description{
To see a list of all methods available for a specific class,
  e.g. the staRVe_model class, run \code{methods(class="staRVe_model")}.
}
