% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/staRVe_model.R
\name{staRVe_model}
\alias{staRVe_model}
\alias{initialize,staRVe_model-method}
\title{Create a \code{staRVe_model} object}
\usage{
staRVe_model(x, ...)

\S4method{initialize}{staRVe_model}(
  .Object,
  process = new("staRVe_process"),
  observations = new("staRVe_observations"),
  settings = new("staRVe_settings")
)
}
\description{
Create a \code{staRVe_model} object
}
\details{
The \code{initialize} function is not mean to be used by the user,
  use \code{staRVe_model} instead.
}
