% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/fit.R, R/staRVe_fit.R
\docType{class}
\name{staRVe_model_fit-class}
\alias{staRVe_model_fit-class}
\alias{staRVe_model_fit}
\alias{staRVe_predict,staRVe_model_fit,sf-method}
\alias{staRVe_predict,staRVe_model_fit,RasterLayer-method}
\alias{tracing,staRVe_model_fit-method}
\alias{TMB_out,staRVe_model_fit-method}
\alias{convergence,staRVe_model_fit-method}
\alias{timing,staRVe_model_fit-method}
\title{An S4 class to hold an optimized staRVe model, extends the staRVe_model class}
\usage{
\S4method{staRVe_predict}{staRVe_model_fit,sf}(x, locations, covariates, time = "model", ...)

\S4method{staRVe_predict}{staRVe_model_fit,RasterLayer}(x, locations, covariates, ...)

\S4method{tracing}{staRVe_model_fit}(x)

\S4method{TMB_out}{staRVe_model_fit}(x)

\S4method{convergence}{staRVe_model_fit}(x)

\S4method{timing}{staRVe_model_fit}(x)
}
\description{
An S4 class to hold an optimized staRVe model, extends the staRVe_model class
}
\section{Methods (by generic)}{
\itemize{
\item \code{staRVe_predict}: Predict/forecast at specific locations

\item \code{staRVe_predict}: Predict/forecast over an entire raster

\item \code{tracing}: Get tracing information

\item \code{TMB_out}: Get TMB objects

\item \code{convergence}: Get convergence message

\item \code{timing}: Get timing information
}}

\section{Slots}{

\describe{
\item{\code{tracing}}{A staRVe_tracing object.}

\item{\code{TMB_out}}{A TMB_out object.}
}}

