% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/staRVe_observation_parameters.R
\name{staRVe_observation_parameters}
\alias{staRVe_observation_parameters}
\alias{initialize,staRVe_observation_parameters-method}
\title{Create a \code{staRVe_observation_parameters} object}
\usage{
staRVe_observation_parameters(x, ...)

\S4method{initialize}{staRVe_observation_parameters}(
  .Object,
  response_distribution = character(0),
  response_parameters = data.frame(par = numeric(0), fixed = numeric(0)),
  link_function = character(0),
  fixed_effects = data.frame(par = numeric(0), fixed = numeric(0))
)
}
\description{
Create a \code{staRVe_observation_parameters} object
}
\details{
The \code{initialize} function is not mean to be used by the user,
  use \code{staRVe_observation_parameters} instead.
}
