% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/staRVe_observations.R
\name{staRVe_observations}
\alias{staRVe_observations}
\alias{initialize,staRVe_observations-method}
\title{Create a \code{staRVe_observations} object}
\usage{
staRVe_observations(x, ...)

\S4method{initialize}{staRVe_observations}(
  .Object,
  data = sf::st_sf(data.frame(y = numeric(1), time = numeric(1)), geometry =
    sf::st_sfc(sf::st_point())),
  transient_graph = new("dag"),
  parameters = new("staRVe_observation_parameters")
)
}
\description{
Create a \code{staRVe_observations} object
}
\details{
The \code{initialize} function is not mean to be used by the user,
  use \code{staRVe_observations} instead.
}
