% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{staRVe_parameters-class}
\alias{staRVe_parameters-class}
\title{An S4 class to collect process and observations parameters.}
\description{
An S4 class to collect process and observations parameters.
}
