% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{staRVe_predict,staRVe_fit,sf-method}
\alias{staRVe_predict,staRVe_fit,sf-method}
\title{Predict at a set of point locations.}
\usage{
\S4method{staRVe_predict}{staRVe_fit,sf}(x, locations, covariates, time = "model", ...)
}
\description{
Predict at a set of point locations.
}
