% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{staRVe_predict}
\alias{staRVe_predict}
\title{Predict from a \code{staRVe_fit} object.}
\usage{
staRVe_predict(x, locations, covariates, time = "model", ...)
}
\arguments{
\item{x}{An object of class \code{staRVe_fit}.}

\item{locations}{Either an object of class \code{sf} containing point geometries,
or an object of class \code{RasterLayer}. If an \code{sf} object, covariate
values and a time index should be included as part of the object.
If a \code{RasterLayer} object, predictions will be made for all raster cells
whose value are not NA. If the raster has no values, then predictions will
be made at every cell. Raster predictions are made at the midpoint of each cell.}

\item{covariates}{A list of \code{Raster*} objects for raster predictions.
 If the model has no covariates, then nothing needs to be supplied.

 If \code{locations} is of class \code{RasterLayer}, then \code{covariates}
 should be a list of \code{Raster*} objects. Each \code{Raster*} object should
 contain data for one covariate, should have one layer for each time unit,
 and should have the same raster geometry as the \code{locations} object. The
 layer names of each raster layer should be of the form \code{T####}, where
 \code{####} gives the specific time index. The geometry of all the
 \code{Raster*} objects should be identical.

 #'  If \code{locations} is of class \code{sf} with point geometries, then
 covariate values should be included in that \code{sf} object.}

\item{time}{What time indices should predictions be made for raster prediction?
If set to "model", predictions are made for every time present in the model.}

\item{...}{Extra options}
}
\value{
Either a \code{sf} object or a list of \code{Raster*} objects,
 containing predictions (with standard errors) for the spatial field, and the linear
 predictor on the link and response scales.
 The return type is the same as the type of input for \code{locations}.
}
\description{
Predict from a \code{staRVe_fit} object.
}
