% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/staRVe_predictions.R
\docType{class}
\name{staRVe_predictions-class}
\alias{staRVe_predictions-class}
\alias{predictions,staRVe_predictions-method}
\alias{locations,staRVe_predictions-method}
\title{An S4 class to hold predictions from a staRVe model.}
\usage{
\S4method{predictions}{staRVe_predictions}(x)

\S4method{locations}{staRVe_predictions}(x)
}
\arguments{
\item{x}{An object}
}
\description{
An S4 class to hold predictions from a staRVe model.
}
\section{Methods (by generic)}{
\itemize{
\item \code{predictions}: Get predictions

\item \code{locations}: Get locations
}}

\section{Slots}{

\describe{
\item{\code{predictions}}{A stars object}

\item{\code{locations}}{An sf object with covariates, a time column, and point geometries.}
}}

