% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/staRVe_process.R
\name{staRVe_process}
\alias{staRVe_process}
\alias{initialize,staRVe_process-method}
\title{Create a \code{staRVe_process} object}
\usage{
staRVe_process(x, ...)

\S4method{initialize}{staRVe_process}(
  .Object,
  random_effects = sf::st_sf(data.frame(w = numeric(1), fixed = logical(1), time =
    numeric(1)), geometry = sf::st_sfc(sf::st_point())),
  persistent_graph = new("dag"),
  parameters = new("staRVe_process_parameters")
)
}
\description{
Create a \code{staRVe_process} object
}
\details{
The \code{initialize} function is not mean to be used by the user,
  use \code{staRVe_process} instead.
}
