% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/staRVe_process_parameters.R
\name{staRVe_process_parameters}
\alias{staRVe_process_parameters}
\alias{initialize,staRVe_process_parameters-method}
\title{Create a \code{staRVe_process_parameters} object}
\usage{
staRVe_process_parameters(x, ...)

\S4method{initialize}{staRVe_process_parameters}(
  .Object,
  covariance_function = character(0),
  spatial_parameters = data.frame(par = numeric(0), fixed = numeric(0)),
  time_parameters = data.frame(par = numeric(0), fixed = numeric(0))
)
}
\description{
Create a \code{staRVe_process_parameters} object
}
\details{
The \code{initialize} function is not mean to be used by the user,
  use \code{staRVe_process_parameters} instead.
}
