% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/staRVe_settings.R
\docType{class}
\name{staRVe_settings-class}
\alias{staRVe_settings-class}
\alias{formula,staRVe_settings-method}
\alias{n_neighbours,staRVe_settings-method}
\alias{p_far_neighbours,staRVe_settings-method}
\alias{distance_units,staRVe_settings-method}
\alias{max_distance,staRVe_settings-method}
\title{An S4 class to hold extra settings for a staRVe model}
\usage{
\S4method{formula}{staRVe_settings}(x)

\S4method{n_neighbours}{staRVe_settings}(x)

\S4method{p_far_neighbours}{staRVe_settings}(x)

\S4method{distance_units}{staRVe_settings}(x)

\S4method{max_distance}{staRVe_settings}(x)
}
\description{
An S4 class to hold extra settings for a staRVe model
}
\section{Methods (by generic)}{
\itemize{
\item \code{formula}: Get/set model formula

\item \code{n_neighbours}: Get/set maximum number of neighbours

\item \code{p_far_neighbours}: Get/set percentage of far neighbours

\item \code{distance_units}: Get/set distance units

\item \code{max_distance}: Get/set maximum distance for neighbours
}}

\section{Slots}{

\describe{
\item{\code{formula}}{A formula}

\item{\code{n_neighbours}}{The number of parents for each node in the graph.}

\item{\code{p_far_neighbours}}{The proportion of parents randomly chosen for each node.}

\item{\code{distance_units}}{The units used for distance calculation.  Must be compatible
with units::set_units.}

\item{\code{max_distance}}{The maximum distance to look for parents. Does not affect
parents coming from p_far_neighbours.}
}}

