% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/staRVe_settings.R
\name{staRVe_settings}
\alias{staRVe_settings}
\alias{initialize,staRVe_settings-method}
\title{Create a \code{staRVe_settings} object}
\usage{
staRVe_settings(x, ...)

\S4method{initialize}{staRVe_settings}(
  .Object,
  formula = new("formula"),
  n_neighbours = 10,
  p_far_neighbours = 0.2,
  distance_units = "km",
  max_distance = Inf
)
}
\description{
Create a \code{staRVe_settings} object
}
\details{
The \code{initialize} function is not mean to be used by the user,
  use \code{staRVe_settings} instead.
}
