% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{staRVe_tracing}
\alias{staRVe_tracing}
\title{An S4 class to hold optimization tracing for a staRVe model}
\description{
An S4 class to hold optimization tracing for a staRVe model
}
\section{Slots}{

\describe{
\item{\code{opt_time}}{A proc_time object. Time used to compute the ML estimates.}

\item{\code{hess_time}}{A proc_time object. Time used to compute the hessian matrix.}

\item{\code{sdr_time}}{A proc_time object. Time used to compute standard errors.}

\item{\code{parameter_hessian}}{The hessian matrix for parameter estimates.}

\item{\code{parameter_covariance}}{The covariance matrix for parameter estimates.}
}}

