% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/fit.R
\name{strv_predict}
\alias{strv_predict}
\alias{strv_predict,starve_fit,sf-method}
\alias{strv_predict,starve_fit,RasterLayer-method}
\title{Use a fitted model to predict.}
\usage{
strv_predict(x, new_data, ...)

\S4method{strv_predict}{starve_fit,sf}(x, new_data)

\S4method{strv_predict}{starve_fit,RasterLayer}(x, new_data, covariates, time = "model")
}
\arguments{
\item{x}{A fitted model}

\item{new_data}{The new set of data for which to predict}

\item{...}{Extra options}

\item{covariates}{A list of \code{Raster*} objects for raster predictions.
 If the model has no covariates, then nothing needs to be supplied.

 If \code{new_data} is of class \code{RasterLayer}, then \code{covariates}
 should be a list of \code{Raster*} objects. Each \code{Raster*} object should
 contain data for one covariate, should have one layer for each time unit,
 and should have the same raster geometry as the \code{new_data} object. The
 layer names of each raster layer should be of the form \code{T####}, where
 \code{####} gives the specific time index. The geometry of all the
 \code{Raster*} objects should be identical.}

\item{time}{Integer vector. At what time indices should predictions be made?
 For the default value "model", predictions are made for every time present
in the model data.}
}
\value{
Predictions for the new data
}
\description{
Use a fitted model to predict.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = starve_fit,new_data = sf}: Predict/forecast at the specific locations and
times given in \code{new_data}. Any covariates used to fit the model
should be included in the rows of \code{new_data}. Returns a \code{long_stars}
object containing a copy of \code{new_data} and the associated predictions
and standard errors for the random effects and response mean on both the
link and natural scale.

\item \code{x = starve_fit,new_data = RasterLayer}: Predictions will be made for all raster cells
whose value are not NA. If the raster has no values, then predictions will
be made at every raster cell. Raster predictions are not treated as areal
data, instead point predictions are made at the midpoint of each raster cell.
The value of the midpoint prediction is taken as the prediction for that cell.
Returns a \code{stars} object whose first two dimensions are the raster
geometry of \code{new_data}, the third dimension is the time index given
in \code{time}, and the fourth dimension is the response variable.
}}

