% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staRVe_observations.R
\name{strv_prepare_observations}
\alias{strv_prepare_observations}
\title{Prepare the observation part of a starve object.}
\usage{
strv_prepare_observations(
  data,
  process,
  settings = new("settings"),
  distribution = "gaussian",
  link = "default"
)
}
\arguments{
\item{data}{An sf object containing the observations and covariates.}

\item{process}{A process object.}

\item{settings}{A settings object.}

\item{distribution}{The response distribution to use. Must be one given by
get_starve_distributions("distribution").}

\item{link}{The link function to use. must be one given by
get_starve_distributions("link").}
}
\value{
An observations object.
}
\description{
Creates a new observation object with the correct dimensions
  for the random effect predictions. Initializes the response distribution
  and fixed effect parameters for the model according to the options specified
  in the formula element of the settings argument. Also adds a column "graph_idx"
  to the supplied data.
}
\seealso{
observations_class

strv_prepare_process, strv_prepare
}
\keyword{internal}
