% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/fit.R
\name{strv_simulate}
\alias{strv_simulate}
\alias{strv_simulate,starve-method}
\title{Simulate from a model object.}
\usage{
strv_simulate(object, ...)

\S4method{strv_simulate}{starve}(object, conditional = FALSE, ...)
}
\arguments{
\item{object}{A model object to simulate from.}

\item{...}{Extra options}

\item{conditional}{Logical. If false (default), new random effects and new
observations are simulated. If true, new observations are simulated
conditional on the random effect values in the supplied model.}
}
\value{
A single copy of \code{object} with a simulated dataset replacing
  the original data
}
\description{
Simulate from a model object.
}
\section{Methods (by class)}{
\itemize{
\item \code{starve}: Simulate a new dataset from the model, with the
option to simulate a new set of random effects as well. The parameter values
used in the simulations are those set in \code{parameters(model)}. Returns
a \code{starve} object with simulated random effects (if \code{conditional=FALSE})
and simulated data.
}}

