% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/staRVe_model.R
\name{strv_update}
\alias{strv_update}
\alias{strv_update,starve,TMB_out-method}
\title{Update a model from a fitted object}
\usage{
strv_update(x, y)

\S4method{strv_update}{starve,TMB_out}(x, y)
}
\arguments{
\item{x}{The model to be updated}

\item{y}{The object to update the model with.}
}
\value{
An updated copy of x
}
\description{
Update a model from a fitted object
}
\section{Methods (by class)}{
\itemize{
\item \code{x = starve,y = TMB_out}: Update model parameters and random effects
from a fitted TMB::MakeADFun object
}}

\keyword{internal}
