% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staRVe_model.R
\name{TMB_in,staRVe_model-method}
\alias{TMB_in,staRVe_model-method}
\title{Convert a staRVe_model object to a form suitable for TMB input.}
\usage{
\S4method{TMB_in}{staRVe_model}(x)
}
\description{
Convert a staRVe_model object to a form suitable for TMB input.
}
