% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/TMB_out.R, R/show.R
\docType{class}
\name{TMB_out-class}
\alias{TMB_out-class}
\alias{obj,TMB_out-method}
\alias{opt,TMB_out-method}
\alias{sdr,TMB_out-method}
\alias{convergence,TMB_out-method}
\alias{show,TMB_out-method}
\title{An S4 class to hold TMB-related objects}
\usage{
\S4method{obj}{TMB_out}(x)

\S4method{opt}{TMB_out}(x)

\S4method{sdr}{TMB_out}(x)

\S4method{convergence}{TMB_out}(x)

\S4method{show}{TMB_out}(object)
}
\arguments{
\item{x}{An object}

\item{object}{An object}
}
\description{
An S4 class to hold TMB-related objects
}
\section{Methods (by generic)}{
\itemize{
\item \code{obj}: Get TMB::MakeADFun object

\item \code{opt}: Get output of optimizer

\item \code{sdr}: Get output of TMB::sdreport

\item \code{convergence}: Get convergence message

\item \code{show}: Print method
}}

\section{Slots}{

\describe{
\item{\code{obj}}{The output of TMB::MakeADFun.}

\item{\code{opt}}{The output of nlminb.}

\item{\code{sdr}}{The output of TMB::sdreport.}
}}

