% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_TMB_out.R
\name{TMB_out}
\alias{TMB_out}
\alias{initialize,TMB_out-method}
\title{Create or extract a \code{TMB_out} object.}
\usage{
TMB_out(x, ...)

\S4method{initialize}{TMB_out}(
  .Object,
  obj = list(),
  opt = list(),
  sdr = structure(numeric(0), class = "sdreport"),
  symbolicAnalysis = logical(0),
  TMB_in = list()
)
}
\arguments{
\item{x}{An object to coerce to clas \code{TMB_out}.}
}
\description{
Create or extract a \code{TMB_out} object.
}
\details{
The \code{initeialize} function is not meant to be used by the user,
 use \code{TMB_out} instead.
}
