% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/TMB_out.R, R/show.R
\docType{class}
\name{TMB_out-class}
\alias{TMB_out-class}
\alias{obj,TMB_out-method}
\alias{obj<-,TMB_out-method}
\alias{opt,TMB_out-method}
\alias{opt<-,TMB_out-method}
\alias{sdr,TMB_out-method}
\alias{sdr<-,TMB_out-method}
\alias{convergence,TMB_out-method}
\alias{show,TMB_out-method}
\title{An S4 class to hold TMB-related objects}
\usage{
\S4method{obj}{TMB_out}(x)

\S4method{obj}{TMB_out}(x) <- value

\S4method{opt}{TMB_out}(x)

\S4method{opt}{TMB_out}(x) <- value

\S4method{sdr}{TMB_out}(x)

\S4method{sdr}{TMB_out}(x) <- value

\S4method{convergence}{TMB_out}(x)

\S4method{show}{TMB_out}(object)
}
\arguments{
\item{x}{An object}

\item{value}{A replacement value}

\item{object}{An object}
}
\description{
An S4 class to hold TMB-related objects
}
\section{Functions}{
\itemize{
\item \code{obj,TMB_out-method}: Get TMB::MakeADFun object

\item \code{obj<-,TMB_out-method}: Set TMB::MakeADFun object

\item \code{opt,TMB_out-method}: Get output of optimizer

\item \code{opt<-,TMB_out-method}: Set output of optimizer

\item \code{sdr,TMB_out-method}: Get output of TMB::sdreport

\item \code{sdr<-,TMB_out-method}: Set output of TMB::sdreport

\item \code{convergence,TMB_out-method}: Get convergence message from optimizer

\item \code{show,TMB_out-method}: Print method
}}

\section{Slots}{

\describe{
\item{\code{obj}}{The output of TMB::MakeADFun.}

\item{\code{opt}}{The output of nlminb.}

\item{\code{sdr}}{The output of TMB::sdreport.}
}}

\seealso{
Other starve_classes: 
\code{\link{dag-class}},
\code{\link{long_stars-class}},
\code{\link{observation_parameters-class}},
\code{\link{observations-class}},
\code{\link{parameters-class}},
\code{\link{process-class}},
\code{\link{process_parameters-class}},
\code{\link{settings-class}},
\code{\link{starve-class}},
\code{\link{starve_classes}},
\code{\link{tracing-class}}
}
\concept{starve_classes}
\keyword{internal}
