% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag.R
\name{access_dag}
\alias{access_dag}
\title{Get or set slots from an object of class \code{dag}.}
\arguments{
\item{x}{An object of class \code{dag}.}

\item{value}{A replacement value.}
}
\description{
Get or set slots from an object of class \code{dag}.
}
\concept{access_dag}
