% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/dag.R
\name{distances}
\alias{distances}
\alias{distances<-}
\alias{edges}
\alias{edges<-}
\alias{Access_dag}
\title{Get or set slots from an object of class \code{dag}.}
\usage{
distances(x)

distances(x) <- value

edges(x)

edges(x) <- value
}
\arguments{
\item{x}{An object of class \code{dag}.}

\item{value}{A replacement value.}
}
\description{
Get or set slots from an object of class \code{dag}.
}
\concept{Access_dag}
