% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staRVe_fit.R
\name{access_staRVe_fit}
\alias{access_staRVe_fit}
\alias{tracing,staRVe_fit-method}
\alias{tracing<-,staRVe_fit-method}
\alias{TMB_out,staRVe_fit-method}
\alias{TMB_out<-,staRVe_fit-method}
\alias{convergence,staRVe_fit-method}
\alias{timing,staRVe_fit-method}
\title{Get or set slots from an object of class \code{staRVe_fit}.}
\usage{
\S4method{tracing}{staRVe_fit}(x)

\S4method{tracing}{staRVe_fit}(x) <- value

\S4method{TMB_out}{staRVe_fit}(x)

\S4method{TMB_out}{staRVe_fit}(x) <- value

\S4method{convergence}{staRVe_fit}(x)

\S4method{timing}{staRVe_fit}(x)
}
\arguments{
\item{x}{An object of class \code{staRVe_fit}.}

\item{value}{A replacement value}
}
\description{
Get or set slots from an object of class \code{staRVe_fit}.
}
\concept{access_staRVe_fit}
