% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staRVe_model.R
\name{access_staRVe_model}
\alias{access_staRVe_model}
\alias{process,staRVe_model-method}
\alias{process<-,staRVe_model-method}
\alias{observations,staRVe_model-method}
\alias{observations<-,staRVe_model-method}
\alias{settings,staRVe_model-method}
\alias{settings<-,staRVe_model-method}
\alias{parameters,staRVe_model-method}
\alias{parameters<-,staRVe_model,staRVe_parameters-method}
\alias{data,staRVe_model-method}
\alias{data<-,staRVe_model-method}
\alias{random_effects,staRVe_model-method}
\alias{random_effects<-,staRVe_model-method}
\alias{graph,staRVe_model-method}
\title{Get or set slots from an object of class \code{staRVe_model}.}
\usage{
\S4method{process}{staRVe_model}(x)

\S4method{process}{staRVe_model}(x) <- value

\S4method{observations}{staRVe_model}(x)

\S4method{observations}{staRVe_model}(x) <- value

\S4method{settings}{staRVe_model}(x)

\S4method{settings}{staRVe_model}(x) <- value

\S4method{parameters}{staRVe_model}(x)

\S4method{parameters}{staRVe_model,staRVe_parameters}(x) <- value

\S4method{data}{staRVe_model}(x)

\S4method{data}{staRVe_model}(x) <- value

\S4method{random_effects}{staRVe_model}(x)

\S4method{random_effects}{staRVe_model}(x) <- value

\S4method{graph}{staRVe_model}(x)
}
\arguments{
\item{x}{An object of class \code{staRVe_model}.}

\item{value}{A replacement value}
}
\description{
Get or set slots from an object of class \code{staRVe_model}.
}
\concept{access_staRVe_model}
