% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staRVe_observation_parameters.R
\name{access_staRVe_observation_parameters}
\alias{access_staRVe_observation_parameters}
\alias{response_distribution,staRVe_observation_parameters-method}
\alias{response_distribution<-,staRVe_observation_parameters-method}
\alias{response_parameters,staRVe_observation_parameters-method}
\alias{response_parameters<-,staRVe_observation_parameters-method}
\alias{link_function,staRVe_observation_parameters-method}
\alias{link_function<-,staRVe_observation_parameters-method}
\alias{fixed_effects,staRVe_observation_parameters-method}
\alias{fixed_effects<-,staRVe_observation_parameters-method}
\title{Get or set slots from an object of class \code{staRVe_observation_parameters}.}
\usage{
\S4method{response_distribution}{staRVe_observation_parameters}(x)

\S4method{response_distribution}{staRVe_observation_parameters}(x) <- value

\S4method{response_parameters}{staRVe_observation_parameters}(x)

\S4method{response_parameters}{staRVe_observation_parameters}(x) <- value

\S4method{link_function}{staRVe_observation_parameters}(x)

\S4method{link_function}{staRVe_observation_parameters}(x) <- value

\S4method{fixed_effects}{staRVe_observation_parameters}(x)

\S4method{fixed_effects}{staRVe_observation_parameters}(x) <- value
}
\arguments{
\item{x}{An object of class \code{staRVe_observation_parameters}.}

\item{value}{A replacement value}
}
\description{
Get or set slots from an object of class \code{staRVe_observation_parameters}.
}
\concept{access_staRVe_observation_parameters}
