% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/staRVe_observations.R
\name{data}
\alias{data}
\alias{data<-}
\alias{transient_graph}
\alias{transient_graph<-}
\alias{Access_staRVe_observations}
\title{Get or set slots from an object of class \code{staRVe_observations}.}
\usage{
data(x)

data(x) <- value

transient_graph(x)

transient_graph(x) <- value
}
\arguments{
\item{x}{An object of class \code{staRVe_observations}.}

\item{value}{A replacement value}
}
\description{
Get or set slots from an object of class \code{staRVe_observations}.
}
\concept{Access_staRVe_observations}
