% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staRVe_process.R
\name{access_staRVe_process}
\alias{access_staRVe_process}
\alias{random_effects,staRVe_process-method}
\alias{random_effects<-,staRVe_process-method}
\alias{persistent_graph,staRVe_process-method}
\alias{persistent_graph<-,staRVe_process-method}
\alias{parameters,staRVe_process-method}
\alias{parameters<-,staRVe_process,ANY-method}
\title{Get or set slots from an object of class \code{staRVe_process}.}
\usage{
\S4method{random_effects}{staRVe_process}(x)

\S4method{random_effects}{staRVe_process}(x) <- value

\S4method{persistent_graph}{staRVe_process}(x)

\S4method{persistent_graph}{staRVe_process}(x) <- value

\S4method{parameters}{staRVe_process}(x)

\S4method{parameters}{staRVe_process,ANY}(x) <- value
}
\arguments{
\item{x}{An object of class \code{staRVe_process}.}

\item{value}{A replacement value}
}
\description{
Get or set slots from an object of class \code{staRVe_process}.
}
\concept{access_staRVe_process}
