% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/staRVe_process_parameters.R
\name{covariance_function}
\alias{covariance_function}
\alias{covariance_function<-}
\alias{spatial_parameters}
\alias{spatial_parameters<-}
\alias{time_parameters}
\alias{time_parameters<-}
\alias{Access_staRVe_process_parameters}
\title{Get or set slots from an object of class \code{staRVe_process_parameters}.}
\usage{
covariance_function(x)

covariance_function(x) <- value

spatial_parameters(x)

spatial_parameters(x) <- value

time_parameters(x)

time_parameters(x) <- value
}
\arguments{
\item{x}{An object of class \code{staRVe_process_parameters}.}

\item{value}{A replacement value}
}
\description{
Get or set slots from an object of class \code{staRVe_process_parameters}.
}
\concept{Access_staRVe_process_parameters}
