% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staRVe_settings.R
\name{access_staRVe_settings}
\alias{access_staRVe_settings}
\alias{formula,staRVe_settings-method}
\title{Get or set slots from an object of class \code{staRVe_settings}.}
\usage{
\S4method{formula}{staRVe_settings}(x)

\S4method{formula}{staRVe_settings}(x) <- value

\S4method{n_neighbours}{staRVe_settings}(x)

\S4method{n_neighbours}{staRVe_settings}(x) <- value

\S4method{p_far_neighbours}{staRVe_settings}(x)

\S4method{p_far_neighbours}{staRVe_settings}(x) <- value

\S4method{distance_units}{staRVe_settings}(x)

\S4method{distance_units}{staRVe_settings}(x) <- value

\S4method{max_distance}{staRVe_settings}(x)

\S4method{max_distance}{staRVe_settings}(x) <- value
}
\arguments{
\item{x}{An object of class \code{staRVe_settings}.}

\item{value}{A replacement value}
}
\description{
Get or set slots from an object of class \code{staRVe_settings}.
}
\concept{access_staRVe_settings}
