% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/staRVe_tracing.R
\name{hess_time}
\alias{hess_time}
\alias{hess_time<-}
\alias{opt_time}
\alias{opt_time<-}
\alias{parameter_covariance}
\alias{parameter_covariance<-}
\alias{parameter_hessian}
\alias{parameter_hessian<-}
\alias{sdr_time}
\alias{sdr_time<-}
\alias{timing}
\alias{Access_staRVe_tracing}
\title{Get or set slots from an object of class \code{staRVe_tracing}.}
\usage{
hess_time(x)

hess_time(x) <- value

opt_time(x)

opt_time(x) <- value

parameter_covariance(x)

parameter_covariance(x) <- value

parameter_hessian(x)

parameter_hessian(x) <- value

sdr_time(x)

sdr_time(x) <- value

timing(x)
}
\arguments{
\item{x}{An object of class \code{staRVe_tracing}.}

\item{value}{A replacement value}
}
\description{
Get or set slots from an object of class \code{staRVe_tracing}.
}
\concept{Access_staRVe_tracing}
