% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/TMB_out.R
\name{obj}
\alias{obj}
\alias{obj<-}
\alias{opt}
\alias{opt<-}
\alias{report}
\alias{sdr}
\alias{sdr<-}
\alias{sdreport}
\alias{symbolicAnalysis}
\alias{symbolicAnalysis<-}
\alias{TMB_in}
\alias{Access_TMB_out}
\title{Get or set slots from an object of class \code{TMB_out}.}
\usage{
obj(x)

obj(x) <- value

opt(x)

opt(x) <- value

report(x)

sdr(x)

sdr(x) <- value

sdreport(x)

symbolicAnalysis(x)

symbolicAnalysis(x) <- value

TMB_in(x)
}
\arguments{
\item{x}{An object of class \code{TMB_out}.}

\item{value}{A replacement value}
}
\description{
Get or set slots from an object of class \code{TMB_out}.
}
\concept{Access_TMB_out}
