% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_TMB_out.R
\name{Access_TMB_out}
\alias{Access_TMB_out}
\alias{obj}
\alias{obj<-}
\alias{opt}
\alias{opt<-}
\alias{sdr}
\alias{sdr<-}
\alias{symbolicAnalysis}
\alias{symbolicAnalysis<-}
\alias{TMB_in}
\alias{TMB_in<-}
\alias{report}
\alias{sdreport}
\alias{get_working_pars}
\alias{get_pars}
\alias{get_parameters}
\alias{get_parameters,TMB_out-method}
\alias{get_geo_vars}
\alias{get_geo_vars,TMB_out-method}
\alias{get_proc}
\alias{get_proc,TMB_out-method}
\alias{get_obs}
\alias{get_obs,TMB_out-method}
\alias{get_resp}
\alias{get_resp,TMB_out-method}
\alias{get_observation}
\alias{get_observation,TMB_out-method}
\alias{get_process}
\alias{get_process,TMB_out-method}
\title{Get or set slots from an object of class \code{TMB_out}.}
\usage{
obj(x)

obj(x) <- value

opt(x)

opt(x) <- value

sdr(x)

sdr(x) <- value

symbolicAnalysis(x)

symbolicAnalysis(x) <- value

TMB_in(x)

TMB_in(x) <- value

report(x)

sdreport(x)

get_working_pars(x)

get_pars(x)

get_parameters(x)

\S4method{get_parameters}{TMB_out}(x)

get_geo_vars(x, var, sf_obj, get_sd)

\S4method{get_geo_vars}{TMB_out}(x, var, sf_obj, get_sd = T)

get_proc(x, sf_obj)

\S4method{get_proc}{TMB_out}(x, sf_obj)

get_obs(x, sf_obj)

\S4method{get_obs}{TMB_out}(x, sf_obj)

get_resp(x, sf_obj)

\S4method{get_resp}{TMB_out}(x, sf_obj)

get_observation(x, sf_obj)

\S4method{get_observation}{TMB_out}(x, sf_obj)

get_process(x, sf_obj)

\S4method{get_process}{TMB_out}(x, sf_obj)
}
\arguments{
\item{x}{An obejct of class \code{TMB_out}.}

\item{value}{A replacement value.}
}
\description{
Get or set slots from an object of class \code{TMB_out}.
}
\concept{Access_TMB_out}
