% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMB_out.R
\name{access_TMB_out}
\alias{access_TMB_out}
\alias{obj,TMB_out-method}
\alias{obj<-,TMB_out-method}
\alias{opt,TMB_out-method}
\alias{opt<-,TMB_out-method}
\alias{sdr,TMB_out-method}
\alias{sdr<-,TMB_out-method}
\alias{convergence,TMB_out-method}
\title{Get or set slots from an object of class \code{TMB_out}.}
\usage{
\S4method{obj}{TMB_out}(x)

\S4method{obj}{TMB_out}(x) <- value

\S4method{opt}{TMB_out}(x)

\S4method{opt}{TMB_out}(x) <- value

\S4method{sdr}{TMB_out}(x)

\S4method{sdr}{TMB_out}(x) <- value

\S4method{convergence}{TMB_out}(x)
}
\arguments{
\item{x}{An object of class \code{TMB_out}.}

\item{value}{A replacement value}
}
\description{
Get or set slots from an object of class \code{TMB_out}.
}
\concept{access_TMB_out}
