% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag.R
\name{construct_dag}
\alias{construct_dag}
\title{Construct a directed acyclic graph for an \code{sf} object with point geometry.}
\usage{
construct_dag(x, settings = new("staRVe_settings"), silent = T)
}
\arguments{
\item{x}{An \code{sf} object containing point geometries, which will be the
nodes of the graph.}

\item{settings}{An object of class \code{staRVe_settings}.}

\item{silent}{Should intermediate calculations be shown?}
}
\value{
An object of class \code{dag}.
}
\description{
The directed acyclic graph is stored as an ordered list. The i'th element of
 the list is a vector containing indices j such that there is a directed edge
 from node j to node i. In other words, each node j is a directed parent of node i.
}
\details{
There can only be an edge from vertex i to vertex j if i comes before j in
 the rows of \code{x}.
}
