% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag.R
\name{construct_dag}
\alias{construct_dag}
\title{Construct a directed acyclic graph for an \code{sf} object with point geometry.}
\usage{
construct_dag(x, n_neighbours = 10, max_dist = Inf,
  distance_units = "km", silent = T)
}
\description{
The directed acyclic graph is stored as an ordered list. The i'th element of
 the list is a vector containing indices j such that there is a directed edge
 from node j to node i. In other words, each node j is a directed parent of node i.
}
\details{
There can only be an edge from vertex i to vertex j if i comes before j in
 the rows of \code{x}.
}
