% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag.R
\name{construct_dag}
\alias{construct_dag}
\alias{construct_obs_dag}
\title{Construct directed acyclic graphs}
\usage{
construct_dag(x, settings = new("staRVe_settings"), silent = T)

construct_obs_dag(
  x,
  y,
  time = 0,
  settings = new("staRVe_settings"),
  check_intersection = T,
  silent = T
)
}
\arguments{
\item{x}{An \code{sf} object of point geometries}

\item{settings}{An object of class \code{staRVe_settings}}

\item{silent}{Should intermediate calculations be shown?}

\item{y}{An \code{sf} object of point geometries containing possible parents}

\item{check_intersection}{Logical. If true, duplicated locations will be checked}

\item{method}{What method should be used to construct the graph? If "standard",
  the parents are taken to be the nearest neighbours according to geographic
  distance.

  If "mesh", additional \code{adjacency_matrix} and \code{distance_matrix}
  arguments need to be supplied.
  The 3 nearest neighbours according to geographic distance are taken to be
  the initial parents, then additional parents are added by traversing the edges
  of the supplied graph. The distance between a location and its parents is the
  sum of the path lengths along the graph from the location to that parent.}

\item{adjacency_matrix}{An adjacency matrix for the locations y.}

\item{distance_matrix}{A distance matrix giving the length of the shortest
paths to/from each node in y, following the edges given by adjacency_matrix}
}
\value{
An object of class \code{dag}
}
\description{
A directed acyclic graph is stored as an ordered list. The ith element of the
  list is a vector containing indices j such that there is a directed edge
  from node j to node i, or possibly node (i+k-1) as detailed below.
}
\section{Functions}{
\itemize{
\item \code{construct_dag}: Construct a directed acyclic graph from a single
\code{sf} object. The first element of the ordered list contains the indices
for the first k rows, where k is the n_neighbours setting. The ith element
contains the parents for the (i+k-1)th row of x.

\item \code{construct_obs_dag}: Construct a directed acyclic graph from one \code{sf}
object to another. The i'th element of the list is a vector containing
indices j such that there is a directed edge from node j (in y) to node i (in x).
}}

