% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag.R
\name{construct_dag}
\alias{construct_dag}
\alias{construct_transient_dag}
\title{Construct directed acyclic graphs}
\usage{
construct_dag(x, settings = new("staRVe_settings"), silent = TRUE)

construct_transient_dag(
  x,
  y,
  time = 0,
  settings = new("staRVe_settings"),
  silent = TRUE
)
}
\arguments{
\item{x}{An \code{sf} object of point geometries}

\item{settings}{An object of class \code{staRVe_settings}}

\item{silent}{Should intermediate calculations be shown?}

\item{y}{An \code{sf} object of point geometries containing possible parents}

\item{time}{The time index for each location in x}

\item{check_intersection}{Logical. If true, duplicated locations will be checked}
}
\value{
An object of class \code{dag}
}
\description{
A directed acyclic graph is stored as an ordered list. The ith element of the
  list is a vector containing indices j such that there is a directed edge
  from node j to node i, or possibly node (i+k-1) as detailed below.
}
\section{Functions}{
\itemize{
\item \code{construct_dag}: Construct a directed acyclic graph from a single
\code{sf} object. The first element of the ordered list contains the indices
for the first k rows, where k is the n_neighbours settings. The ith element
contains the parents for the (i+k-1)th row of x. Also returns a sorted copy
of the locations x.

\item \code{construct_transient_dag}: Construct a directed acyclic graph from one \code{sf}
object to another. Output is a list whose t'th element is the transient graph
for time t. y should be the persistent graph locations, x should be data locations.
}}

