% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/dag.R, R/show.R
\docType{class}
\name{dag-class}
\alias{dag-class}
\alias{edges,dag-method}
\alias{distances,dag-method}
\alias{distance_units,dag-method}
\alias{distance_units<-,dag-method}
\alias{show,dag-method}
\title{An S4 class to hold a directed acyclic graph with distances.}
\usage{
\S4method{edges}{dag}(x)

\S4method{distances}{dag}(x)

\S4method{distance_units}{dag}(x)

\S4method{distance_units}{dag}(x) <- value

\S4method{show}{dag}(object)
}
\arguments{
\item{x}{An object}

\item{value}{A replacement value}

\item{object}{An object}
}
\description{
An S4 class to hold a directed acyclic graph with distances.
}
\section{Methods (by generic)}{
\itemize{
\item \code{edges}: Get edge list

\item \code{distances}: Get list of edge distances

\item \code{distance_units}: Get distance units

\item \code{distance_units<-}: Set distance units. Distances are automatically converted
to the new units.

\item \code{show}: Print method
}}

\section{Slots}{

\describe{
\item{\code{edges}}{A list of integer vector, indexing the parents of each node.}

\item{\code{distances}}{A list of matrices containing the distances between
nodes and parents.}

\item{\code{distance_units}}{The units used for distance calculation.  Must be compatible
with units::set_units.}
}}

