% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/dag.R
\name{dag}
\alias{dag}
\alias{initialize,dag-method}
\title{Create a \code{dag} object}
\usage{
dag(x, ...)

\S4method{initialize}{dag}(.Object, edges = list(), distances = list(), distance_units = "km")
}
\description{
Create a \code{dag} object
}
\details{
The \code{initialize} function is not mean to be used by the user,
  use \code{dag} instead.
}
