% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/dag.R, R/show.R
\docType{class}
\name{dag-class}
\alias{dag-class}
\alias{edges,dag-method}
\alias{distances,dag-method}
\alias{distance_units,dag-method}
\alias{distance_units<-,dag-method}
\alias{show,dag-method}
\title{An S4 class to hold a directed acyclic graph with distances.}
\usage{
\S4method{edges}{dag}(x)

\S4method{distances}{dag}(x)

\S4method{distance_units}{dag}(x)

\S4method{distance_units}{dag}(x) <- value

\S4method{show}{dag}(object)
}
\arguments{
\item{x}{An object}

\item{value}{A replacement value}

\item{object}{An object}
}
\description{
An S4 class to hold a directed acyclic graph with distances.
}
\section{Functions}{
\itemize{
\item \code{edges,dag-method}: Get edge list

\item \code{distances,dag-method}: Get list of edge distances

\item \code{distance_units,dag-method}: Get distance units

\item \code{distance_units<-,dag-method}: Set distance units. Edge distances are automatically
converted to the new units.

\item \code{show,dag-method}: Print method
}}

\section{Slots}{

\describe{
\item{\code{edges}}{A list whose elements consist of a "to" vector and a
"from" vector. There is a directed edge in the graph from each vertex in
"from" to each vertex in "to". If there are multiple vertices in "to" then
there are undirected edges between all vertices in "to" (forming a Kn graph).}

\item{\code{distances}}{A list whose elements consist of a distance matrix for the edges
in the correponding entry of edges. The rows/columns of the distance matrix
are in the same order as the vertices in that edge list, starting with the
vertices in "to" and then the vertices in "from".}

\item{\code{distance_units}}{The units used for distance calculation.  Must be compatible
with units::set_units.}
}}

\seealso{
construct_graph Functions used to create graphs.

Other starve_classes: 
\code{\link{TMB_out-class}},
\code{\link{long_stars-class}},
\code{\link{observation_parameters-class}},
\code{\link{observations-class}},
\code{\link{parameters-class}},
\code{\link{process-class}},
\code{\link{process_parameters-class}},
\code{\link{settings-class}},
\code{\link{starve-class}},
\code{\link{tracing-class}}
}
\concept{starve_classes}
