% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag.R
\name{.get_one_dag_node}
\alias{.get_one_dag_node}
\title{Get parent nodes for a node of a directed acyclic graph.}
\usage{
.get_one_dag_node(x, nodes, n_neighbours, max_dist, distance_units,
  silent = T, ...)
}
\arguments{
\item{x}{An \code{sf} object containing one point geometry (if more than one
is present, only the first will be used).}

\item{nodes}{An \code{sf} object containing point geometries. These will
be the possible parent nodes of \code{x}.}

\item{n_neighbours}{A positive integer giving the number of parents for each node.}

\item{max_dist}{The maximum distance (in meters) to search for parents.}

\item{distance_units}{Which units should be used for distances?}
}
\value{
A list with elements "parents" and "dists". The "parents" element
  is a vector giving the indices of the rows of \code{nodes} which are the
  parents of \code{x}. The "dists" element is a matrix whose first row is the
  distance between \code{x} and its parents, and the rest of the matrix gives
  the distances between parent nodes.
}
\description{
Get parent nodes for a node of a directed acyclic graph.
}
