% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_to_input.R
\name{.sf_to_process}
\alias{.sf_to_process}
\title{Convert an sf object to a spatio-temporal Process, with TMB input.}
\usage{
.sf_to_process(sf_object, n_neighbours, time_range, time_type, time_name,
  silent = T, max_dist = Inf, distance_units = "km", ...)
}
\arguments{
\item{sf_object}{An object of class `sf` containing point geometries. All
locations will be used for every time index.}

\item{n_neighbours}{A positive integer giving the (maximum) number of
parents for each node.}

\item{time_type}{Which time structure should be used? Must be one of
\code{independent}, \code{ar1} (default), or \code{rw}.}

\item{silent}{Should intermediate calculations be printed?}

\item{max_dist}{The maximum distance used to search for parents.
Unless this has a units attribute, units are assumed to be the same as
the supplied \code{distance_units}. See \code{\link{construct_dag}}.}

\item{distance_units}{Which units should be used for distances?
See \code{\link{construct_dag}}.}

\item{n_time}{A positive integer giving the number of time indices.}
}
\value{
A list with 4 elements:
  \describe{
    \item{Process}{An sf object containing point geometries for each random
      effect and each time point.}
    \item{data}{A list to pass to TMB::MakeADFun.}
    \item{pars}{A list to pass to TMB::MakeADFun.}
    \item{rand}{A list to pass to TMB::MakeADFun.}
    \item{map}{A list to pass to TMB::MakeADFun.}
 }
}
\description{
Convert an sf object to a spatio-temporal Process, with TMB input.
}
