% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_to_input.R
\name{get_staRVe_distributions}
\alias{get_staRVe_distributions}
\title{Print a list of implemented response distributions and link functions.}
\usage{
get_staRVe_distributions(which = c("distribution", "link", "covariance"))
}
\arguments{
\item{which}{A character vector containing "distribution", "link", and/or
"covariance". Defaults to all.}
}
\value{
A named character vector giving the implemented distributions,
 link functions, and/or covariance functions, depending on the value of the
 `which` parameter.
}
\description{
Print a list of implemented response distributions and link functions.
}
