% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{haddock_survey}
\alias{haddock_survey}
\title{Haddock survey data from the Scotian shelf}
\format{
An sf object of point geometries containing 6,434 rows and 11 fields:
\describe{
  \item{TotalWeight}{Total biomass caught}
  \item{TotalCount}{Total number of indidivuals caught}
  \item{Presence}{Presence (1) or absence (0) of fish}
  \item{Year}{The year of the sample}
  \item{Duration}{The duration of the fishing set}
  \item{Distance}{The distance traveled during the fishing set}
  \item{Speed}{The average speed traveled during the fishing set}
  \item{SurfaceTemperature}{Sea surface temperature}
  \item{BottomTemperature}{Sea bottom temperature}
  \item{BottomSalinity}{Water salinity at the seafloor}
  \item{Area}{The area swept by the fishing set}
}
}
\usage{
haddock_survey
}
\description{
A spatio-temporal dataset containing observations of haddock abundance
  from a yearly scientific survey on the Scotian shelf from the years 1982 to 2017.
}
\details{
The data were obtained from the Department of Fisheries and Oceans Canada.
}
\keyword{datasets}
