% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{idx_exchange}
\alias{idx_exchange}
\alias{idxC_to_R}
\alias{idxR_to_C}
\title{Exchange indices between C++ and R}
\usage{
idxC_to_R(x)

idxR_to_C(x)
}
\arguments{
\item{x}{An object with numeric entries}
}
\description{
Converts indices to be used in C++ (where indices start at 0) and R
  (where indices start at 1).
}
\section{Functions}{
\itemize{
\item \code{idxC_to_R}: Add 1 to all indices to convert from C++ to R

\item \code{idxR_to_C}: Subtract 1 from all indices to convert from R to C++
}}

\keyword{internal}
