% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/long_stars.R
\docType{class}
\name{long_stars-class}
\alias{long_stars-class}
\alias{values,long_stars-method}
\alias{values<-,long_stars-method}
\alias{locations,long_stars-method}
\alias{locations<-,long_stars-method}
\title{An S4 class to hold predictions from a staRVe model.}
\usage{
\S4method{values}{long_stars}(x)

\S4method{values}{long_stars}(x) <- value

\S4method{locations}{long_stars}(x)

\S4method{locations}{long_stars}(x) <- value
}
\arguments{
\item{x}{An object}

\item{value}{A replacement value}
}
\description{
An S4 class to hold predictions from a staRVe model.
}
\section{Methods (by generic)}{
\itemize{
\item \code{values}: Get predictions

\item \code{values<-}: Set predictions

\item \code{locations}: Get locations

\item \code{locations<-}: Set locations
}}

\section{Slots}{

\describe{
\item{\code{predictions}}{A stars object}

\item{\code{locations}}{An sf object with covariates, a time column, and point geometries.}
}}

