% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\docType{methods}
\name{predict_staRVe}
\alias{predict_staRVe}
\alias{predict_staRVe,staRVe,sf-method}
\alias{predict_staRVe,staRVe,RasterLayer-method}
\title{Predict from a \code{staRVe} object.}
\usage{
predict_staRVe(x, locations, ...)

\S4method{predict_staRVe}{staRVe,sf}(x, locations, covariates,
  n_neighbours = "model", max_dist = "model", time = "model", ...)

\S4method{predict_staRVe}{staRVe,RasterLayer}(x, locations, covariates,
  ...)
}
\arguments{
\item{x}{An object of class \code{staRVe}.}

\item{locations}{Either an object of class \code{sf} containing point geometries,
or an object of class \code{RasterLayer}. If a \code{RasterLayer} object, predictions
will be made for all raster cells whose value are not NA. If the raster has no values,
then predictions will be made at every cell. This object should not have any
time information.}

\item{covariates}{Either a data.frame of class \code{sf} or a list of \code{Raster*}
 objects, depending on the input type of \code{locations}.

 If \code{locations} is of class \code{sf} with point geometries, then
 \code{covariates} should also be of class \code{sf}. The data.frame should contain
 columns for each of the covariates, a column for the time index, and a point
 geometry. For each time unit and prediction point, there should be a row
 in the \code{covariates} data.frame, but the rows do not need to be in the
 same order as \code{locations}.

 If \code{locations} is of class \code{RasterLayer}, then \code{covariates}
 should be a list of \code{Raster*} objects. Each \code{Raster*} object should
 contain data for one covariate, should have one layer for each time unit,
 and should have the same raster geometry as the \code{locations} object. The
 layer names of each raster layer should be of the form \code{T####}, where
 \code{####} gives the specific time unit.}

\item{n_neighbours}{The maximum number of parents each node should have.
If set to "model", then this number is taken from the model directed acyclic graph.}

\item{max_dist}{The maximum distance to search for parents. If set to "model,
the maximum distance is set to the two times the spatial range parameter.
Unless this has a units attribute, units are assumed to be the same as the model.
See \code{\link{construct_dag}}.}

\item{time}{What time units should predictions be made for? If set to "model",
predictions are made for every time present in the model. The supplied time
units must be a subset of those in the model.}

\item{runOrderings}{A logical value. Should TMB::runSymbolicAnalysis be used?}

\item{get_sd}{Should standard errors be computed?
Warning: computing standard errors can take a significant amount of time.}

\item{parallel_cores}{How many cores to run in parallel? Defaults to one.
See parallel::detectCores() for the number of cores on your machine. Uses
parallel::mclapply, and thus never runs in parallel on Windows machines.}
}
\value{
Either a \code{sf} object or a list of \code{Raster*} objects,
 containing predictions (with standard errors) for the spatial field and for
 the response. The return type is the same as the type of input for \code{locations}.
}
\description{
Predict from a \code{staRVe} object.
}
