% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staRVe_model.R
\name{prepare_staRVe_model}
\alias{prepare_staRVe_model}
\title{Create an object of class \code{staRVe_model}.}
\usage{
prepare_staRVe_model(
  formula,
  data,
  nodes = data,
  n_neighbours = 10,
  p_far_neighbours = 0.2,
  distribution = "gaussian",
  link = "identity",
  silent = T,
  max_dist = Inf,
  distance_units = "km",
  fit = F,
  ...
)
}
\arguments{
\item{formula}{A formula object used to describe the model. The details are given
under `Details'.}

\item{data}{An object of class `sf` containing point geometries. Data
used for the `formula' object will be found here.}

\item{nodes}{An object of class `sf` containing point geometries.
The default value uses the same locations as the observations. These
locations will be used as the nodes for the random effects. All locations
will be used for each year.}

\item{n_neighbours}{An integer giving the number of parents for each node.}

\item{p_far_neighbours}{What percent of neighbours should be randomly selected?}

\item{distribution}{A character vector giving the response distribution. The
default is "gaussian". The "atLeastOneBinomial" distribution models the probability of at least one success in n trials. See \code{get_staRVe_distributions("distribution")}.}

\item{link}{A character vector giving the response link function. The default
is "identity". See \code{get_staRVe_distributions("link")}.}

\item{silent}{Should intermediate calculations be printed?}

\item{max_dist}{The maximum distance used to search for parents.
Unless this has a units attribute, units are assumed to be the same as
the supplied \code{distance_units}.}

\item{distance_units}{Which units should be used for distances?}

\item{fit}{Should the model be fit in this call? If true, returns a fitted model.}
}
\value{
A staRVe_model object. If fit=T, a staRVe_fit object.
}
\description{
.
}
\details{
The formula object should always be of the form
  \code{y ~ sample.size(n)+mean(x+z) + time(t,type="ar1") + space("matern",nu=1.5)}.
The variable y should be replaced with the desired response variable, and t
should be replaced with the desired time index. There are currently three
valid options for the `type' argument in \code{time(t,type="ar1")} -- "ar1" for
an AR(1) structure, "rw" for a random walk, and "independent" for independent
spatial fields each year. The \code{space(...)} term specifies the spatial covariance
function.

The sample.size(...) term is only used if the response distribution is \code{binomial} or \code{atLeastOneBinomial}. If it is missing the sample sizes are assumed to all be 1.

If the \code{time(...)} term is missing, all observations are assumed to be
at the same time. If the \code{space(...)} term is missing, the spatial
covariance function defaults to the exponential function.

The variables in the \code{mean(...)} ``special" are linear predictors for the mean
of the response variable. Any formula valid for the \code{lm} command can be used
inside the \code{mean(...)}, however any missing values will likely cause errors.
}
