% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staRVe_observations.R
\name{prepare_staRVe_observations}
\alias{prepare_staRVe_observations}
\title{Prepare the observation part of a staRVe_model object.}
\usage{
prepare_staRVe_observations(
  data,
  process,
  settings = new("staRVe_settings"),
  distribution = "gaussian",
  link = "identity"
)
}
\arguments{
\item{data}{An sf object containing the observations and covariates.}

\item{process}{A staRVe_process object.}

\item{settings}{A staRVe_settings object.}

\item{distribution}{The response distribution to use. must be one given by
get_staRVe_distributions("distribution").}

\item{link}{The link function to use. must be one given by
get_staRVe_distributions("link").}
}
\value{
A staRVe_observations object.
}
\description{
Prepare the observation part of a staRVe_model object.
}
