% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staRVe_process.R
\name{prepare_staRVe_process}
\alias{prepare_staRVe_process}
\title{Prepare the process part of a staRVe_model object.}
\usage{
prepare_staRVe_process(
  data,
  nodes = data,
  persistent_graph = NA,
  settings = new("staRVe_settings")
)
}
\arguments{
\item{data}{An sf object containing the observation locations and times.}

\item{nodes}{An sf object containing the persistent graph locations (defaults to data)}

\item{persistent_graph}{An optionally supplied pre-computed persistent graph.
If a persistent_graph is not supplied, the order of the rows of nodes may change.
If a persisten_graph is supplied, the order will not change. See \link{construct_dag}.}

\item{settings}{A staRVe_settings object}
}
\value{
A staRVe_process object
}
\description{
Creates a new staRVe_process object with the correct dimensions for the
  temporal random effects, persistent graph random effects, and transient
  graph random effects. Initializes the temporal and spatial parameters for
  the model according to the options specified in the formula element of the
  settings argument. Constructs the persistent and transient graph.
}
\details{
Before creating the persistent graph any duplicate locations in nodes
  are removed. Before creating the transient graph any location in data that
  is present in nodes is removed
}
\seealso{
staRVe_process

prepare_staRVe_observations, prepare_staRVe_model
}
\keyword{internal}
