% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/staRVe_process.R
\docType{class}
\name{process-class}
\alias{process-class}
\alias{time_effects,process-method}
\alias{time_effects<-,process-method}
\alias{pg_re,process-method}
\alias{pg_re<-,process-method}
\alias{tg_re,process-method}
\alias{tg_re<-,process-method}
\alias{persistent_graph,process-method}
\alias{persistent_graph<-,process-method}
\alias{transient_graph,process-method}
\alias{transient_graph<-,process-method}
\alias{parameters,process-method}
\alias{parameters<-,process,ANY-method}
\alias{random_effects,process-method}
\title{An S4 class to hold the process information for a starve model.}
\usage{
\S4method{time_effects}{process}(x)

\S4method{time_effects}{process}(x) <- value

\S4method{pg_re}{process}(x)

\S4method{pg_re}{process}(x) <- value

\S4method{tg_re}{process}(x)

\S4method{tg_re}{process}(x) <- value

\S4method{persistent_graph}{process}(x)

\S4method{persistent_graph}{process}(x) <- value

\S4method{transient_graph}{process}(x)

\S4method{transient_graph}{process}(x) <- value

\S4method{parameters}{process}(x)

\S4method{parameters}{process,ANY}(x) <- value

\S4method{random_effects}{process}(x)
}
\arguments{
\item{x}{An object}

\item{value}{A replacement value}
}
\description{
An S4 class to hold the process information for a starve model.
}
\section{Functions}{
\itemize{
\item \code{time_effects,process-method}: Get temporal random effects

\item \code{time_effects<-,process-method}: Set temporal random effects

\item \code{pg_re,process-method}: Get spatio-temporal random effects for persistent graph

\item \code{pg_re<-,process-method}: Set spatio-temporal random effects for persistent graph

\item \code{tg_re,process-method}: Get spatio-temporal random effects for transient graph

\item \code{tg_re<-,process-method}: Set spatio-temporal random effects for transient graph

\item \code{persistent_graph,process-method}: Get persistent graph

\item \code{persistent_graph<-,process-method}: Set persistent graph (for internal use only)

\item \code{transient_graph,process-method}: Get transient graph

\item \code{transient_graph<-,process-method}: Set transient graph (for internal use  only)

\item \code{parameters,process-method}: Get parameters as a process_parameters object

\item \code{parameters<-,process,ANY-method}: Set parameters using a new process_parameters
object. Not the recommended way to modify specific parameter values, instead see the
package vignette.

\item \code{random_effects,process-method}: Get list containing the persistent graph random
effects and the transient graph random effects.
}}

\section{Slots}{

\describe{
\item{\code{time_effects}}{A stars object containing temporal random effects}

\item{\code{random_effects}}{A stars object containing spatio-temporal random effects.}

\item{\code{persistent_graph}}{A dag object describing the dependence graph of the process.}

\item{\code{parameters}}{An object of class process_parameters.}
}}

\seealso{
strv_prepare_process

Other starve_classes: 
\code{\link{TMB_out-class}},
\code{\link{dag-class}},
\code{\link{long_stars-class}},
\code{\link{observation_parameters-class}},
\code{\link{observations-class}},
\code{\link{parameters-class}},
\code{\link{process_parameters-class}},
\code{\link{settings-class}},
\code{\link{starve-class}},
\code{\link{starve_classes}},
\code{\link{tracing-class}}
}
\concept{starve_classes}
\keyword{internal}
