% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/staRVe_model.R, R/show.R
\docType{class}
\name{starve-class}
\alias{starve-class}
\alias{process,starve-method}
\alias{process<-,starve-method}
\alias{observations,starve-method}
\alias{observations<-,starve-method}
\alias{settings,starve-method}
\alias{settings<-,starve-method}
\alias{tracing,starve-method}
\alias{tracing<-,starve-method}
\alias{TMB_out,starve-method}
\alias{TMB_out<-,starve-method}
\alias{time_effects,starve-method}
\alias{time_effects<-,starve-method}
\alias{pg_re,starve-method}
\alias{pg_re<-,starve-method}
\alias{tg_re,starve-method}
\alias{tg_re<-,starve-method}
\alias{random_effects,starve-method}
\alias{persistent_graph,starve-method}
\alias{persistent_graph<-,starve-method}
\alias{transient_graph,starve-method}
\alias{transient_graph<-,starve-method}
\alias{graph,starve-method}
\alias{covariance_function,starve-method}
\alias{covariance_function<-,starve-method}
\alias{spatial_parameters,starve-method}
\alias{spatial_parameters<-,starve-method}
\alias{time_parameters,starve-method}
\alias{time_parameters<-,starve-method}
\alias{dat,starve-method}
\alias{dat<-,starve-method}
\alias{data_predictions,starve-method}
\alias{data_predictions<-,starve-method}
\alias{response_distribution,starve-method}
\alias{response_distribution<-,starve-method}
\alias{response_parameters,starve-method}
\alias{response_parameters<-,starve-method}
\alias{link_function,starve-method}
\alias{link_function<-,starve-method}
\alias{fixed_effects,starve-method}
\alias{fixed_effects<-,starve-method}
\alias{distance_units,starve-method}
\alias{distance_units<-,starve-method}
\alias{formula,starve-method}
\alias{formula<-,starve-method}
\alias{time_name,starve-method}
\alias{parameters,starve-method}
\alias{parameters<-,starve,parameters-method}
\alias{convergence,starve-method}
\alias{timing,starve-method}
\alias{show,starve-method}
\title{An S4 class describing a starve model.}
\usage{
\S4method{process}{starve}(x)

\S4method{process}{starve}(x) <- value

\S4method{observations}{starve}(x)

\S4method{observations}{starve}(x) <- value

\S4method{settings}{starve}(x)

\S4method{settings}{starve}(x) <- value

\S4method{tracing}{starve}(x)

\S4method{tracing}{starve}(x) <- value

\S4method{TMB_out}{starve}(x)

\S4method{TMB_out}{starve}(x) <- value

\S4method{time_effects}{starve}(x)

\S4method{time_effects}{starve}(x) <- value

\S4method{pg_re}{starve}(x)

\S4method{pg_re}{starve}(x) <- value

\S4method{tg_re}{starve}(x)

\S4method{tg_re}{starve}(x) <- value

\S4method{random_effects}{starve}(x)

\S4method{persistent_graph}{starve}(x)

\S4method{persistent_graph}{starve}(x) <- value

\S4method{transient_graph}{starve}(x)

\S4method{transient_graph}{starve}(x) <- value

\S4method{graph}{starve}(x)

\S4method{covariance_function}{starve}(x)

\S4method{covariance_function}{starve}(x) <- value

\S4method{spatial_parameters}{starve}(x)

\S4method{spatial_parameters}{starve}(x) <- value

\S4method{time_parameters}{starve}(x)

\S4method{time_parameters}{starve}(x) <- value

\S4method{dat}{starve}(x)

\S4method{dat}{starve}(x) <- value

\S4method{data_predictions}{starve}(x)

\S4method{data_predictions}{starve}(x) <- value

\S4method{response_distribution}{starve}(x)

\S4method{response_distribution}{starve}(x) <- value

\S4method{response_parameters}{starve}(x)

\S4method{response_parameters}{starve}(x) <- value

\S4method{link_function}{starve}(x)

\S4method{link_function}{starve}(x) <- value

\S4method{fixed_effects}{starve}(x)

\S4method{fixed_effects}{starve}(x) <- value

\S4method{distance_units}{starve}(x)

\S4method{distance_units}{starve}(x) <- value

\S4method{formula}{starve}(x)

\S4method{formula}{starve}(x) <- value

\S4method{time_name}{starve}(x)

\S4method{parameters}{starve}(x)

\S4method{parameters}{starve,parameters}(x) <- value

\S4method{convergence}{starve}(x)

\S4method{timing}{starve}(x)

\S4method{show}{starve}(object)
}
\arguments{
\item{x}{An object}

\item{value}{A replacement value}

\item{object}{An object}
}
\description{
An S4 class describing a starve model.
}
\section{Functions}{
\itemize{
\item \code{process,starve-method}: Get the process part of the model (for internal use only)

\item \code{process<-,starve-method}: Set the process part of the model (for internal use only)

\item \code{observations,starve-method}: Get the observation part of the model (for internal use only)

\item \code{observations<-,starve-method}: Set the observation part of the model (for internal use only)

\item \code{settings,starve-method}: Get model settings

\item \code{settings<-,starve-method}: Set model settings (for internal use only)

\item \code{tracing,starve-method}: Get tracing information, see \link{tracing_class}.

\item \code{tracing<-,starve-method}: Set tracing information (for internal use only)

\item \code{TMB_out,starve-method}: Get TMB objects (for internal use only)

\item \code{TMB_out<-,starve-method}: Set TMB objects (for internal use only)

\item \code{time_effects,starve-method}: Get temporal random effects

\item \code{time_effects<-,starve-method}: Set temporal random effects

\item \code{pg_re,starve-method}: Get persistent graph random effects

\item \code{pg_re<-,starve-method}: Set spatio-temporal random effects

\item \code{tg_re,starve-method}: Get transient graph random effects

\item \code{tg_re<-,starve-method}: Set transient graph random effects

\item \code{random_effects,starve-method}: Get list containing the persistent graph random
effects and the transient graph random effects.

\item \code{persistent_graph,starve-method}: Get persistent graph

\item \code{persistent_graph<-,starve-method}: Set persistent graph (for internal use only)

\item \code{transient_graph,starve-method}: Get transient graph

\item \code{transient_graph<-,starve-method}: Set transient graph (for internal use only)

\item \code{graph,starve-method}: Get a list containing the persistent and transient graphs

\item \code{covariance_function,starve-method}: Get covariance function

\item \code{covariance_function<-,starve-method}: Set covariance function(s). Run
get_starve_distributions("covariance") for valid covariance functions.
Setting the covariance function also overwrites the spatial parameters.

\item \code{spatial_parameters,starve-method}: Get spatial parameters

\item \code{spatial_parameters<-,starve-method}: Set spatial parameters

\item \code{time_parameters,starve-method}: Get time parameters

\item \code{time_parameters<-,starve-method}: Set time parameters

\item \code{dat,starve-method}: Get data including response variables, time indices,
locations, covariates, etc.

\item \code{dat<-,starve-method}: Set data. Warning: if you want to add new
rows you must give each additional row a value in the graph_idx column,
or create a new object from scratch.

\item \code{data_predictions,starve-method}: Get data predictions, a long_stars object
with the data (see \code{dat}) and associated random effect
predictions.

\item \code{data_predictions<-,starve-method}: Set data predictions

\item \code{response_distribution,starve-method}: Get the response distribution(s)

\item \code{response_distribution<-,starve-method}: Set the response distribution(s). Run
get_starve_distributions("distribution") for valid options.
Setting the response distribution also overwrites the response
parameters and link function(s).

\item \code{response_parameters,starve-method}: Get response distribution parameters

\item \code{response_parameters<-,starve-method}: Set response distribution parameters

\item \code{link_function,starve-method}: Get link function(s)

\item \code{link_function<-,starve-method}: Set link function(s). Run
get_starve_distributions("link") for valid link functions.

\item \code{fixed_effects,starve-method}: Get fixed effect parameters

\item \code{fixed_effects<-,starve-method}: Set fixed effect parameters

\item \code{distance_units,starve-method}: Get distance units used for the model

\item \code{distance_units<-,starve-method}: Set distance units used for the model

\item \code{formula,starve-method}: Get formula used for model

\item \code{formula<-,starve-method}: Set formula used for the model

\item \code{time_name,starve-method}: Get the name of the time variable used in
the model formula (for internal use only)

\item \code{parameters,starve-method}: Get all model parameters as a parameters object

\item \code{parameters<-,starve,parameters-method}: Set all model parameters with a new
parameters object

\item \code{convergence,starve-method}: Get convergence message from optimizer

\item \code{timing,starve-method}: Get all timing information, see \link{tracing}

\item \code{show,starve-method}: Print method
}}

\section{Slots}{

\describe{
\item{\code{process}}{A process object.}

\item{\code{observations}}{An observations object.}

\item{\code{settings}}{A settings object.}
}}

\seealso{
strv_prepare

Other starve_classes: 
\code{\link{TMB_out-class}},
\code{\link{dag-class}},
\code{\link{long_stars-class}},
\code{\link{observation_parameters-class}},
\code{\link{observations-class}},
\code{\link{parameters-class}},
\code{\link{process-class}},
\code{\link{process_parameters-class}},
\code{\link{settings-class}},
\code{\link{starve_classes}},
\code{\link{tracing-class}}
}
\concept{starve_classes}
