% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{staRVe_classes}
\alias{staRVe_classes}
\title{A list of classes in the staRVe package.}
\description{
To see documentation for a particular class, load the package and run e.g.
  class?staRVe_model. To see a list of all methods available for a specific class,
  see the help page or run \code{methods(class="staRVe_model")}.
}
\section{Classes}{

\itemize{
  \item dag
  \item long_stars
  \item staRVe_process_parameters
  \item staRVe_process
  \item staRVe_observation_parameters
  \item staRVe_observations
  \item staRVe_settings
  \item staRVe_model
  \item staRVe_tracing
  \item TMB_out
  \item staRVe_model_fit
  \item staRVe_parameters
}
}

