% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{starve_classes}
\alias{starve_classes}
\title{A list of classes in the starve package.}
\description{
To see documentation for a particular class, load the package and run e.g.
  class?starve. To see a list of all methods available for a specific class,
  see the help page for that class or run e.g. \code{methods(class="starve")}.
}
\section{Classes}{

\itemize{
  \item dag
  \item long_stars
  \item settings
  \item process_parameters
  \item process
  \item observation_parameters
  \item observations
  \item tracing
  \item TMB_out
  \item starve
  \item parameters
}
}

\seealso{
Other starve_classes: 
\code{\link{TMB_out-class}},
\code{\link{dag-class}},
\code{\link{long_stars-class}},
\code{\link{observation_parameters-class}},
\code{\link{observations-class}},
\code{\link{parameters-class}},
\code{\link{process-class}},
\code{\link{process_parameters-class}},
\code{\link{settings-class}},
\code{\link{starve-class}},
\code{\link{tracing-class}}
}
\concept{starve_classes}
\keyword{internal}
