% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/staRVe_fit.R
\docType{class}
\name{staRVe_fit-class}
\alias{staRVe_fit-class}
\alias{tracing,staRVe_fit-method}
\alias{TMB_out,staRVe_fit-method}
\alias{convergence,staRVe_fit-method}
\alias{timing,staRVe_fit-method}
\title{An S4 class to hold an optimized staRVe model, extends the staRVe_model class}
\usage{
\S4method{tracing}{staRVe_fit}(x)

\S4method{TMB_out}{staRVe_fit}(x)

\S4method{convergence}{staRVe_fit}(x)

\S4method{timing}{staRVe_fit}(x)
}
\description{
An S4 class to hold an optimized staRVe model, extends the staRVe_model class
}
\section{Methods (by generic)}{
\itemize{
\item \code{tracing}: Get tracing information

\item \code{TMB_out}: Get TMB objects

\item \code{convergence}: Get convergence message

\item \code{timing}: Get timing information
}}

\section{Slots}{

\describe{
\item{\code{tracing}}{A staRVe_tracing object.}

\item{\code{TMB_out}}{A TMB_out object.}
}}

