% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{staRVe_fit}
\alias{staRVe_fit}
\title{Fit a \code{staRVe_model} object.}
\usage{
staRVe_fit(x, ...)
}
\arguments{
\item{x}{A staRVe_model object.}

\item{silent}{Should tracing information be printed?}
}
\value{
A staRVe_fit object.
}
\description{
After creating a \code{staRVe_model} object using \code{prepare_staRVe_model},
  this function is used to find maximum likelihood estimates of the parameters
  and random effects.
}
