% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/fit.R
\name{staRVe_fit}
\alias{staRVe_fit}
\alias{staRVe_fit,staRVe_model-method}
\title{Fit a model object.}
\usage{
staRVe_fit(x, ...)

\S4method{staRVe_fit}{staRVe_model}(x, silent = FALSE, ...)
}
\arguments{
\item{x}{A model object to be fitted.}

\item{...}{Options to be passed}

\item{silent}{Should intermediate calculations be printed?}
}
\value{
A fitted model object.
}
\description{
Takes an unfitted model object and performs inference.
}
\section{Methods (by class)}{
\itemize{
\item \code{staRVe_model}: Takes an unfitted staRVe_model object and performs
maximum likelihood inference via the \code{TMB} package to obtain parameter
and random effect estimates with their associated standard errors. Options
in the \dots argument are passed to TMB::MakeADFun and TMB::sdreport.
}}

