% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/show.R, R/staRVe_model.R
\docType{class}
\name{staRVe_model-class}
\alias{staRVe_model-class}
\alias{staRVe_model}
\alias{show,staRVe_model-method}
\alias{process,staRVe_model-method}
\alias{process<-,staRVe_model-method}
\alias{observations,staRVe_model-method}
\alias{observations<-,staRVe_model-method}
\alias{settings,staRVe_model-method}
\alias{settings<-,staRVe_model-method}
\alias{time_effects,staRVe_model-method}
\alias{time_effects<-,staRVe_model-method}
\alias{pg_re,staRVe_model-method}
\alias{pg_re<-,staRVe_model-method}
\alias{tg_re,staRVe_model-method}
\alias{tg_re<-,staRVe_model-method}
\alias{random_effects,staRVe_model-method}
\alias{persistent_graph,staRVe_model-method}
\alias{persistent_graph<-,staRVe_model-method}
\alias{transient_graph,staRVe_model-method}
\alias{transient_graph<-,staRVe_model-method}
\alias{graph,staRVe_model-method}
\alias{covariance_function,staRVe_model-method}
\alias{covariance_function<-,staRVe_model-method}
\alias{spatial_parameters,staRVe_model-method}
\alias{spatial_parameters<-,staRVe_model-method}
\alias{time_parameters,staRVe_model-method}
\alias{time_parameters<-,staRVe_model-method}
\alias{dat,staRVe_model-method}
\alias{dat<-,staRVe_model-method}
\alias{data_predictions,staRVe_model-method}
\alias{data_predictions<-,staRVe_model-method}
\alias{response_distribution,staRVe_model-method}
\alias{response_distribution<-,staRVe_model-method}
\alias{response_parameters,staRVe_model-method}
\alias{response_parameters<-,staRVe_model-method}
\alias{link_function,staRVe_model-method}
\alias{link_function<-,staRVe_model-method}
\alias{fixed_effects,staRVe_model-method}
\alias{fixed_effects<-,staRVe_model-method}
\alias{distance_units,staRVe_model-method}
\alias{distance_units<-,staRVe_model-method}
\alias{formula,staRVe_model-method}
\alias{formula<-,staRVe_model-method}
\alias{.time_name,staRVe_model-method}
\alias{parameters,staRVe_model-method}
\alias{parameters<-,staRVe_model,staRVe_parameters-method}
\title{An S4 class describing a staRVe model.}
\usage{
\S4method{show}{staRVe_model}(object)

\S4method{process}{staRVe_model}(x)

\S4method{process}{staRVe_model}(x) <- value

\S4method{observations}{staRVe_model}(x)

\S4method{observations}{staRVe_model}(x) <- value

\S4method{settings}{staRVe_model}(x)

\S4method{settings}{staRVe_model}(x) <- value

\S4method{time_effects}{staRVe_model}(x)

\S4method{time_effects}{staRVe_model}(x) <- value

\S4method{pg_re}{staRVe_model}(x)

\S4method{pg_re}{staRVe_model}(x) <- value

\S4method{tg_re}{staRVe_model}(x)

\S4method{tg_re}{staRVe_model}(x) <- value

\S4method{random_effects}{staRVe_model}(x)

\S4method{persistent_graph}{staRVe_model}(x)

\S4method{persistent_graph}{staRVe_model}(x) <- value

\S4method{transient_graph}{staRVe_model}(x)

\S4method{transient_graph}{staRVe_model}(x) <- value

\S4method{graph}{staRVe_model}(x)

\S4method{covariance_function}{staRVe_model}(x)

\S4method{covariance_function}{staRVe_model}(x) <- value

\S4method{spatial_parameters}{staRVe_model}(x)

\S4method{spatial_parameters}{staRVe_model}(x) <- value

\S4method{time_parameters}{staRVe_model}(x)

\S4method{time_parameters}{staRVe_model}(x) <- value

\S4method{dat}{staRVe_model}(x)

\S4method{dat}{staRVe_model}(x) <- value

\S4method{data_predictions}{staRVe_model}(x)

\S4method{data_predictions}{staRVe_model}(x) <- value

\S4method{response_distribution}{staRVe_model}(x)

\S4method{response_distribution}{staRVe_model}(x) <- value

\S4method{response_parameters}{staRVe_model}(x)

\S4method{response_parameters}{staRVe_model}(x) <- value

\S4method{link_function}{staRVe_model}(x)

\S4method{link_function}{staRVe_model}(x) <- value

\S4method{fixed_effects}{staRVe_model}(x)

\S4method{fixed_effects}{staRVe_model}(x) <- value

\S4method{distance_units}{staRVe_model}(x)

\S4method{distance_units}{staRVe_model}(x) <- value

\S4method{formula}{staRVe_model}(x)

\S4method{formula}{staRVe_model}(x) <- value

\S4method{.time_name}{staRVe_model}(x)

\S4method{parameters}{staRVe_model}(x)

\S4method{parameters}{staRVe_model,staRVe_parameters}(x) <- value
}
\arguments{
\item{object}{An object}

\item{x}{An object}

\item{value}{A replacement value}
}
\description{
An S4 class describing a staRVe model.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Print method

\item \code{process}: Get the process part of the model (for internal use only)

\item \code{process<-}: Set the process part of the model (for internal use only)

\item \code{observations}: Get the observation part of the model (for internal use only)

\item \code{observations<-}: Set the observation part of the model (for internal use only)

\item \code{settings}: Get model settings

\item \code{settings<-}: Set model settings (for internal use only)

\item \code{time_effects}: Get temporal random effects

\item \code{time_effects<-}: Set temporal random effects

\item \code{pg_re}: Get persistent graph random effects

\item \code{pg_re<-}: Set spatio-temporal random effects

\item \code{tg_re}: Get transient graph random effects

\item \code{tg_re<-}: Set transient graph random effects

\item \code{random_effects}: Get list containing the persistent graph random
effects and the transient graph random effects.

\item \code{persistent_graph}: Get persistent graph

\item \code{persistent_graph<-}: Set persistent graph (for internal use only)

\item \code{transient_graph}: Get transient graph

\item \code{transient_graph<-}: Set transient graph (for internal use only)

\item \code{graph}: Get a list containing the persistent and transient graphs

\item \code{covariance_function}: Get covariance function

\item \code{covariance_function<-}: Set covariance function(s). Run
get_staRVe_distributions("covariance") for valid covariance functions.
Setting the covariance function also overwrites the spatial parameters.

\item \code{spatial_parameters}: Get spatial parameters

\item \code{spatial_parameters<-}: Set spatial parameters

\item \code{time_parameters}: Get time parameters

\item \code{time_parameters<-}: Set time parameters

\item \code{dat}: Get data including response variables, time indices,
locations, covariates, etc.

\item \code{dat<-}: Set data. Warning: if you add new rows to the
data.frame you also need to manually update the transient graph,
transient graph random effects, and the graph_idx column for the
new rows.

\item \code{data_predictions}: Get data predictions, a long_stars object
with the data (see \code{dat}) and associated random effect
predictions.

\item \code{data_predictions<-}: Set data predictions

\item \code{response_distribution}: Get the response distribution(s)

\item \code{response_distribution<-}: Set the response distribution(s). Run
get_staRVe_distributions("distribution") for valid options.
Setting the response distribution also overwrites the response
parameters and link function(s).

\item \code{response_parameters}: Get response distribution parameters

\item \code{response_parameters<-}: Set response distribution parameters

\item \code{link_function}: Get link function(s)

\item \code{link_function<-}: Set link function(s). Run
get_staRVe_distributions("link") for valid link functions.

\item \code{fixed_effects}: Get fixed effect parameters

\item \code{fixed_effects<-}: Set fixed effect parameters

\item \code{distance_units}: Get distance units used for the model

\item \code{distance_units<-}: Set distance units used for the model

\item \code{formula}: Get formula used for model

\item \code{formula<-}: Set formula used for the model

\item \code{.time_name}: Get the name of the time variable used in
the model formula (for internal use only)

\item \code{parameters}: Get all model parameters as a staRVe_parameters object

\item \code{parameters<-}: Set all model parameters with a new
staRVe_parameters object
}}

\section{Slots}{

\describe{
\item{\code{process}}{A staRVe_process object.}

\item{\code{observations}}{A staRVe_observations object.}

\item{\code{settings}}{A staRVe_settings object.}
}}

\seealso{
prepare_staRVe_model

Other staRVe_classes: 
\code{\link{TMB_out-class}},
\code{\link{dag-class}},
\code{\link{long_stars-class}},
\code{\link{staRVe_classes}},
\code{\link{staRVe_model_fit-class}},
\code{\link{staRVe_observation_parameters-class}},
\code{\link{staRVe_observations-class}},
\code{\link{staRVe_parameters-class}},
\code{\link{staRVe_process-class}},
\code{\link{staRVe_process_parameters-class}},
\code{\link{staRVe_settings-class}},
\code{\link{staRVe_tracing-class}}
}
\concept{staRVe_classes}
