% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/fit.R, R/show.R, R/staRVe_model.R
\docType{class}
\name{staRVe_model-class}
\alias{staRVe_model-class}
\alias{staRVe_model}
\alias{staRVe_fit,staRVe_model-method}
\alias{staRVe_simulate,staRVe_model-method}
\alias{show,staRVe_model-method}
\alias{settings,staRVe_model-method}
\alias{time_effects,staRVe_model-method}
\alias{time_effects<-,staRVe_model-method}
\alias{random_effects,staRVe_model-method}
\alias{random_effects<-,staRVe_model-method}
\alias{covariance_function,staRVe_model-method}
\alias{covariance_function<-,staRVe_model-method}
\alias{spatial_parameters,staRVe_model-method}
\alias{spatial_parameters<-,staRVe_model-method}
\alias{time_parameters,staRVe_model-method}
\alias{time_parameters<-,staRVe_model-method}
\alias{dat,staRVe_model-method}
\alias{dat<-,staRVe_model-method}
\alias{response_distribution,staRVe_model-method}
\alias{response_distribution<-,staRVe_model-method}
\alias{response_parameters,staRVe_model-method}
\alias{response_parameters<-,staRVe_model-method}
\alias{link_function,staRVe_model-method}
\alias{link_function<-,staRVe_model-method}
\alias{fixed_effects,staRVe_model-method}
\alias{fixed_effects<-,staRVe_model-method}
\alias{distance_units,staRVe_model-method}
\alias{distance_units<-,staRVe_model-method}
\alias{parameters,staRVe_model-method}
\alias{parameters<-,staRVe_model,staRVe_parameters-method}
\alias{graph,staRVe_model-method}
\title{An S4 class describing a staRVe model.}
\usage{
\S4method{staRVe_fit}{staRVe_model}(x, silent = F, ...)

\S4method{staRVe_simulate}{staRVe_model}(model, conditional = F, ...)

\S4method{show}{staRVe_model}(object)

\S4method{settings}{staRVe_model}(x)

\S4method{time_effects}{staRVe_model}(x)

\S4method{time_effects}{staRVe_model}(x) <- value

\S4method{random_effects}{staRVe_model}(x)

\S4method{random_effects}{staRVe_model}(x) <- value

\S4method{covariance_function}{staRVe_model}(x)

\S4method{covariance_function}{staRVe_model}(x) <- value

\S4method{spatial_parameters}{staRVe_model}(x)

\S4method{spatial_parameters}{staRVe_model}(x) <- value

\S4method{time_parameters}{staRVe_model}(x)

\S4method{time_parameters}{staRVe_model}(x) <- value

\S4method{dat}{staRVe_model}(x)

\S4method{dat}{staRVe_model}(x) <- value

\S4method{response_distribution}{staRVe_model}(x)

\S4method{response_distribution}{staRVe_model}(x) <- value

\S4method{response_parameters}{staRVe_model}(x)

\S4method{response_parameters}{staRVe_model}(x) <- value

\S4method{link_function}{staRVe_model}(x)

\S4method{link_function}{staRVe_model}(x) <- value

\S4method{fixed_effects}{staRVe_model}(x)

\S4method{fixed_effects}{staRVe_model}(x) <- value

\S4method{distance_units}{staRVe_model}(x)

\S4method{distance_units}{staRVe_model}(x) <- value

\S4method{parameters}{staRVe_model}(x)

\S4method{parameters}{staRVe_model,staRVe_parameters}(x) <- value

\S4method{graph}{staRVe_model}(x)
}
\arguments{
\item{x}{An object}

\item{silent}{Should intermediate calculations be printed?}

\item{...}{Extra options to supply to TMB::MakeADFun or TMB::sdreport}

\item{model}{A staRVe_model}

\item{conditional}{logical. If true, new observations are simulated conditional
on the random effect values in \code{random_effects(model)}.
If false, new random effects and new observations are simulated.}

\item{object}{An object}

\item{value}{A replacement value}
}
\description{
An S4 class describing a staRVe model.
}
\section{Methods (by generic)}{
\itemize{
\item \code{staRVe_fit}: Find maximum likelihood estimate of parameters
and random effects

\item \code{staRVe_simulate}: Simulate from the model

\item \code{show}: Print method

\item \code{settings}: Get model settings

\item \code{time_effects}: Get temporal random effects

\item \code{time_effects<-}: Set temporal random effects

\item \code{random_effects}: Get spatio-temporal random effects

\item \code{random_effects<-}: Set spatio-temporal random effects

\item \code{covariance_function}: Get covariance function

\item \code{covariance_function<-}: Set covariance function. Run
get_staRVe_distributions("covariance") for valid covariance functions.
Setting the covariance function also overwrites the spatial parameters.

\item \code{spatial_parameters}: Get spatial parameters

\item \code{spatial_parameters<-}: Get spatial parameters

\item \code{time_parameters}: Get time parameters

\item \code{time_parameters<-}: Set time parameters

\item \code{dat}: Get data

\item \code{dat<-}: Set data

\item \code{response_distribution}: Get the response distribution.

\item \code{response_distribution<-}: Set the response distribution. Run
get_staRVe_distributions("distribution") for valid options.
Setting the response distribution also overwrites the response
parameters and link function.

\item \code{response_parameters}: Get response parameters

\item \code{response_parameters<-}: Set response parameters

\item \code{link_function}: Get the link function

\item \code{link_function<-}: Set the link function. Run
get_staRVe_distributions("link") for valid covariance functions.

\item \code{fixed_effects}: Get fixed effects

\item \code{fixed_effects<-}: Set fixed effects

\item \code{distance_units}: Get distance units used for the model

\item \code{distance_units<-}: Set distance units used for the model

\item \code{parameters}: Get parameters

\item \code{parameters<-}: Set parameters

\item \code{graph}: Get a list containing the persistent and transient graphs
}}

\section{Slots}{

\describe{
\item{\code{process}}{A staRVe_process object.}

\item{\code{observations}}{A staRVe_observations object.}

\item{\code{settings}}{A staRVe_settings object.}
}}

