% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{staRVe_model}
\alias{staRVe_model}
\title{An S4 class describing a staRVe model.}
\description{
An S4 class describing a staRVe model.
}
\section{Slots}{

\describe{
\item{\code{A}}{staRVe_process object.}

\item{\code{A}}{staRVe_observations object.}

\item{\code{A}}{staRVe_settings object.}
}}

