% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/fit.R, R/show.R, R/staRVe_fit.R
\docType{class}
\name{staRVe_model_fit-class}
\alias{staRVe_model_fit-class}
\alias{staRVe_model_fit}
\alias{staRVe_predict,staRVe_model_fit,sf-method}
\alias{staRVe_predict,staRVe_model_fit,RasterLayer-method}
\alias{show,staRVe_model_fit-method}
\alias{tracing,staRVe_model_fit-method}
\alias{TMB_out,staRVe_model_fit-method}
\alias{convergence,staRVe_model_fit-method}
\alias{timing,staRVe_model_fit-method}
\title{An S4 class to hold an optimized staRVe model, extends the staRVe_model class}
\usage{
\S4method{staRVe_predict}{staRVe_model_fit,sf}(x, locations)

\S4method{staRVe_predict}{staRVe_model_fit,RasterLayer}(x, locations, covariates, time = "model")

\S4method{show}{staRVe_model_fit}(object)

\S4method{tracing}{staRVe_model_fit}(x)

\S4method{TMB_out}{staRVe_model_fit}(x)

\S4method{convergence}{staRVe_model_fit}(x)

\S4method{timing}{staRVe_model_fit}(x)
}
\arguments{
\item{x}{An object}

\item{locations}{An sf object for prediction locations and times, with covariates}

\item{covariates}{A list of \code{Raster*} objects for raster predictions.
 If the model has no covariates, then nothing needs to be supplied.

 If \code{locations} is of class \code{RasterLayer}, then \code{covariates}
 should be a list of \code{Raster*} objects. Each \code{Raster*} object should
 contain data for one covariate, should have one layer for each time unit,
 and should have the same raster geometry as the \code{locations} object. The
 layer names of each raster layer should be of the form \code{T####}, where
 \code{####} gives the specific time index. The geometry of all the
 \code{Raster*} objects should be identical.

 #'  If \code{locations} is of class \code{sf} with point geometries, then
 covariate values should be included in that \code{sf} object.}

\item{time}{What time indices should predictions be made for raster prediction?
If set to "model", predictions are made for every time present in the model.}

\item{object}{An object}

\item{...}{Extra options}
}
\description{
An S4 class to hold an optimized staRVe model, extends the staRVe_model class
}
\section{Methods (by generic)}{
\itemize{
\item \code{staRVe_predict}: Predict/forecast at specific locations

\item \code{staRVe_predict}: Predict/forecast over an entire raster

\item \code{show}: Print method

\item \code{tracing}: Get tracing information

\item \code{TMB_out}: Get TMB objects

\item \code{convergence}: Get convergence message

\item \code{timing}: Get timing information
}}

\section{Slots}{

\describe{
\item{\code{tracing}}{A staRVe_tracing object.}

\item{\code{TMB_out}}{A TMB_out object.}
}}
