% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/show.R, R/staRVe_model_fit.R
\docType{class}
\name{staRVe_model_fit-class}
\alias{staRVe_model_fit-class}
\alias{staRVe_model_fit}
\alias{show,staRVe_model_fit-method}
\alias{tracing,staRVe_model_fit-method}
\alias{tracing<-,staRVe_model_fit-method}
\alias{TMB_out,staRVe_model_fit-method}
\alias{TMB_out<-,staRVe_model_fit-method}
\alias{convergence,staRVe_model_fit-method}
\alias{timing,staRVe_model_fit-method}
\title{An extension of the staRVe_model class to hold optimization information}
\usage{
\S4method{show}{staRVe_model_fit}(object)

\S4method{tracing}{staRVe_model_fit}(x)

\S4method{tracing}{staRVe_model_fit}(x) <- value

\S4method{TMB_out}{staRVe_model_fit}(x)

\S4method{TMB_out}{staRVe_model_fit}(x) <- value

\S4method{convergence}{staRVe_model_fit}(x)

\S4method{timing}{staRVe_model_fit}(x)
}
\arguments{
\item{object}{An object}

\item{x}{An object}

\item{value}{A replacement value}
}
\description{
An extension of the staRVe_model class to hold optimization information
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Print method

\item \code{tracing}: Get tracing information, see \link{staRVe_tracing}.

\item \code{tracing<-}: Set tracing information (for internal use only)

\item \code{TMB_out}: Get TMB objects (for internal use only)

\item \code{TMB_out<-}: Set TMB objects (for internal use only)

\item \code{convergence}: Get convergence message from optimizer

\item \code{timing}: Get all timing information, see \link{staRVe_tracing}
}}

\section{Slots}{

\describe{
\item{\code{tracing}}{A staRVe_tracing object}

\item{\code{TMB_out}}{A TMB_out object (for internal use)}
}}

\seealso{
staRVe_model, staRVe_tracing, TMB_out

Other staRVe_classes: 
\code{\link{TMB_out-class}},
\code{\link{dag-class}},
\code{\link{long_stars-class}},
\code{\link{staRVe_classes}},
\code{\link{staRVe_model-class}},
\code{\link{staRVe_observation_parameters-class}},
\code{\link{staRVe_observations-class}},
\code{\link{staRVe_parameters-class}},
\code{\link{staRVe_process-class}},
\code{\link{staRVe_process_parameters-class}},
\code{\link{staRVe_settings-class}},
\code{\link{staRVe_tracing-class}}
}
\concept{staRVe_classes}
