% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{staRVe_observation_parameters-class}
\alias{staRVe_observation_parameters-class}
\title{An S4 class to hold observation parameters for a staRVe model.}
\description{
An S4 class to hold observation parameters for a staRVe model.
}
