% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/staRVe_observation_parameters.R
\docType{class}
\name{staRVe_observation_parameters-class}
\alias{staRVe_observation_parameters-class}
\alias{response_distribution,staRVe_observation_parameters-method}
\alias{response_distribution<-,staRVe_observation_parameters-method}
\alias{response_parameters,staRVe_observation_parameters-method}
\alias{response_parameters<-,staRVe_observation_parameters-method}
\alias{link_function,staRVe_observation_parameters-method}
\alias{link_function<-,staRVe_observation_parameters-method}
\alias{fixed_effects,staRVe_observation_parameters-method}
\alias{fixed_effects<-,staRVe_observation_parameters-method}
\title{An S4 class to hold observation parameters for a staRVe model.}
\usage{
\S4method{response_distribution}{staRVe_observation_parameters}(x)

\S4method{response_distribution}{staRVe_observation_parameters}(x) <- value

\S4method{response_parameters}{staRVe_observation_parameters}(x)

\S4method{response_parameters}{staRVe_observation_parameters}(x) <- value

\S4method{link_function}{staRVe_observation_parameters}(x)

\S4method{link_function}{staRVe_observation_parameters}(x) <- value

\S4method{fixed_effects}{staRVe_observation_parameters}(x)

\S4method{fixed_effects}{staRVe_observation_parameters}(x) <- value
}
\arguments{
\item{x}{An object}

\item{value}{A replacement value}
}
\description{
An S4 class to hold observation parameters for a staRVe model.
}
\section{Methods (by generic)}{
\itemize{
\item \code{response_distribution}: Get response distribution(s)

\item \code{response_distribution<-}: Get/set response distribution(s). Run
get_staRVe_distributions("distribution") for valid options. Setting the
the response distribution also overwrites the response parameters and
link function(s).

\item \code{response_parameters}: Get response distribution parameters

\item \code{response_parameters<-}: Set response distribution parameters

\item \code{link_function}: Get link function(s)

\item \code{link_function<-}: set link function(s).  Run
get_staRVe_distributions("link") for valid options.

\item \code{fixed_effects}: Get fixed effect parameters

\item \code{fixed_effects<-}: Set fixed effect parameters
}}

\section{Slots}{

\describe{
\item{\code{response_distribution}}{The response distribution(s), must be one from
get_staRVe_distributions("distribution").}

\item{\code{response_parameters}}{A list containing a data.frame of response distribution
parameters for each response variable.}

\item{\code{link_function}}{The link function(s), must be one from
get_staRVe_distributions("link").}

\item{\code{fixed_effects}}{A list containing a data.frame of fixed effect parameters
for each response variable.}
}}

\seealso{
Other staRVe_classes: 
\code{\link{TMB_out-class}},
\code{\link{dag-class}},
\code{\link{long_stars-class}},
\code{\link{staRVe_classes}},
\code{\link{staRVe_model-class}},
\code{\link{staRVe_model_fit-class}},
\code{\link{staRVe_observations-class}},
\code{\link{staRVe_parameters-class}},
\code{\link{staRVe_process-class}},
\code{\link{staRVe_process_parameters-class}},
\code{\link{staRVe_settings-class}},
\code{\link{staRVe_tracing-class}}
}
\concept{staRVe_classes}
\keyword{internal}
