% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{staRVe_observation_parameters}
\alias{staRVe_observation_parameters}
\title{An S4 class to hold observation parameters for a staRVe model.}
\description{
An S4 class to hold observation parameters for a staRVe model.
}
\section{Slots}{

\describe{
\item{\code{response_distribution}}{The response distribution, must be one given by
get_staRVe_distributions("distribution").}

\item{\code{response_parameters}}{A data.frame containing parameters for the response
distribution.}

\item{\code{link_function}}{The link function, must be one given by
get_staRVe_distributions("link").}

\item{\code{fixed_effects}}{A data.frame containing the fixed effect parameters.}
}}

