% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{staRVe_observations-class}
\alias{staRVe_observations-class}
\title{An S4 class to hold the observation information for a staRVe model.}
\description{
An S4 class to hold the observation information for a staRVe model.
}
\seealso{
Other staRVe-classes: 
\code{\link{staRVe_process-class}}
}
\concept{staRVe-classes}
