% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/staRVe_observations.R
\docType{class}
\name{staRVe_observations-class}
\alias{staRVe_observations-class}
\alias{dat,staRVe_observations-method}
\alias{dat<-,staRVe_observations-method}
\alias{parameters,staRVe_observations-method}
\alias{parameters<-,staRVe_observations,ANY-method}
\title{An S4 class to hold the observation information for a staRVe model.}
\usage{
\S4method{dat}{staRVe_observations}(x)

\S4method{dat}{staRVe_observations}(x) <- value

\S4method{parameters}{staRVe_observations}(x)

\S4method{parameters}{staRVe_observations,ANY}(x) <- value
}
\arguments{
\item{x}{An object}

\item{value}{A replacement value}
}
\description{
An S4 class to hold the observation information for a staRVe model.
}
\section{Methods (by generic)}{
\itemize{
\item \code{dat}: Get data

\item \code{dat<-}: Set data

\item \code{parameters}: Get parameters

\item \code{parameters<-}: Set parameters
}}

\section{Slots}{

\describe{
\item{\code{data}}{An sf object containing the data for the model.}

\item{\code{transient_graph}}{A dag object describing the dependence of the data on
the process.}

\item{\code{parameters}}{An object of class staRVe_observation_parameters.}
}}

