% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{staRVe_observations-class}
\alias{staRVe_observations-class}
\title{An S4 class to hold the observation information for a staRVe model.}
\description{
An S4 class to hold the observation information for a staRVe model.
}
\section{Slots}{

\describe{
\item{\code{data}}{An sf object containing the data for the model.}

\item{\code{transient_graph}}{A dag object describing the dependence of the data on
the process.}

\item{\code{parameters}}{An object of class staRVe_observation_parameters.}
}}

