% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/staRVe_process.R
\docType{class}
\name{staRVe_process-class}
\alias{staRVe_process-class}
\alias{time_effects,staRVe_process-method}
\alias{time_effects<-,staRVe_process-method}
\alias{pg_re,staRVe_process-method}
\alias{pg_re<-,staRVe_process-method}
\alias{tg_re,staRVe_process-method}
\alias{tg_re<-,staRVe_process-method}
\alias{parameters,staRVe_process-method}
\alias{parameters<-,staRVe_process,ANY-method}
\alias{random_effects,staRVe_process-method}
\title{An S4 class to hold the process information for a staRVe model.}
\usage{
\S4method{time_effects}{staRVe_process}(x)

\S4method{time_effects}{staRVe_process}(x) <- value

\S4method{pg_re}{staRVe_process}(x)

\S4method{pg_re}{staRVe_process}(x) <- value

\S4method{tg_re}{staRVe_process}(x)

\S4method{tg_re}{staRVe_process}(x) <- value

\S4method{parameters}{staRVe_process}(x)

\S4method{parameters}{staRVe_process,ANY}(x) <- value

\S4method{random_effects}{staRVe_process}(x)
}
\arguments{
\item{x}{An object}

\item{value}{A replacement value}
}
\description{
An S4 class to hold the process information for a staRVe model.
}
\section{Methods (by generic)}{
\itemize{
\item \code{time_effects}: Get temporal random effects

\item \code{time_effects<-}: Set temporal random effects

\item \code{pg_re}: Get spatio-temporal random effects for persistent graph

\item \code{pg_re<-}: Set spatio-temporal random effects for persistent graph

\item \code{tg_re}: Get spatio-temporal random effects for transient graph

\item \code{tg_re<-}: Set spatio-temporal random effects for transient graph

\item \code{parameters}: Set parameters

\item \code{parameters<-}: Set parameters

\item \code{random_effects}: Get list of random effects
}}

\section{Slots}{

\describe{
\item{\code{time_effects}}{A stars object containing temporal random effects}

\item{\code{random_effects}}{A stars object containing spatio-temporal random effects.}

\item{\code{persistent_graph}}{A dag object describing the dependence graph of the process.}

\item{\code{parameters}}{An object of class staRVe_process_parameters.}
}}

