% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/staRVe_process.R
\docType{class}
\name{staRVe_process-class}
\alias{staRVe_process-class}
\alias{time_effects,staRVe_process-method}
\alias{random_effects,staRVe_process-method}
\alias{parameters,staRVe_process-method}
\title{An S4 class to hold the process information for a staRVe model.}
\usage{
\S4method{time_effects}{staRVe_process}(x)

\S4method{random_effects}{staRVe_process}(x)

\S4method{parameters}{staRVe_process}(x)
}
\description{
An S4 class to hold the process information for a staRVe model.
}
\section{Methods (by generic)}{
\itemize{
\item \code{time_effects}: Get/set temporal random effects

\item \code{random_effects}: Get/set spatio-temporal random effects

\item \code{parameters}: Get/set parameters
}}

\section{Slots}{

\describe{
\item{\code{random_effects}}{An sf object containing random effects.}

\item{\code{persistent}}{graph A dag object describing the dependence graph of the process.}

\item{\code{parameters}}{An object of class staRVe_process_parameters.}
}}

