% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{staRVe_process-class}
\alias{staRVe_process-class}
\title{An S4 class to hold the process information for a staRVe model.}
\description{
An S4 class to hold the process information for a staRVe model.
}
\section{Slots}{

\describe{
\item{\code{random_effects}}{An sf object containing random effects.}

\item{\code{persistent}}{graph A dag object describing the dependence graph of the process.}

\item{\code{parameters}}{An object of class staRVe_process_parameters.}
}}

