% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{staRVe_process_parameters-class}
\alias{staRVe_process_parameters-class}
\title{An S4 class to hold process parameters for a staRVe model.}
\description{
An S4 class to hold process parameters for a staRVe model.
}
\section{Slots}{

\describe{
\item{\code{covariance_function}}{The covariance function, must be one given by
get_staRVe_distributions("covariance").}

\item{\code{spatial_parameters}}{A data.frame containing spatial parameters.}

\item{\code{time_parameters}}{A data.frame containing time parameters.}
}}

