% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{staRVe_settings-class}
\alias{staRVe_settings-class}
\title{An S4 class to hold extra settings for a staRVe model}
\description{
An S4 class to hold extra settings for a staRVe model
}
\section{Slots}{

\describe{
\item{\code{formula}}{A formula}

\item{\code{n_neighbours}}{The number of parents for each node in the graph.}

\item{\code{p_far_neighbours}}{The proportion of parents randomly chosen for each node.}

\item{\code{distance_units}}{The units used for distance calculation.  Must be compatible
with units::set_units.}

\item{\code{max_distance}}{The maximum distance to look for parents. Does not affect
parents coming from p_far_neighbours.}
}}

