% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/staRVe_settings.R
\docType{class}
\name{staRVe_settings-class}
\alias{staRVe_settings-class}
\alias{staRVe_settings}
\alias{formula,staRVe_settings-method}
\alias{formula<-,staRVe_settings-method}
\alias{n_neighbours,staRVe_settings-method}
\alias{p_far_neighbours,staRVe_settings-method}
\alias{distance_units,staRVe_settings-method}
\alias{distance_units<-,staRVe_settings-method}
\alias{max_distance,staRVe_settings-method}
\alias{obs_dag_method,staRVe_settings-method}
\alias{extras,staRVe_settings-method}
\alias{extras<-,staRVe_settings-method}
\title{An S4 class to hold extra settings for a staRVe model}
\usage{
\S4method{formula}{staRVe_settings}(x)

\S4method{formula}{staRVe_settings}(x) <- value

\S4method{n_neighbours}{staRVe_settings}(x)

\S4method{p_far_neighbours}{staRVe_settings}(x)

\S4method{distance_units}{staRVe_settings}(x)

\S4method{distance_units}{staRVe_settings}(x) <- value

\S4method{max_distance}{staRVe_settings}(x)

\S4method{obs_dag_method}{staRVe_settings}(x)

\S4method{extras}{staRVe_settings}(x)

\S4method{extras}{staRVe_settings}(x) <- value
}
\arguments{
\item{x}{An object}

\item{value}{A replacement value}
}
\description{
An S4 class to hold extra settings for a staRVe model
}
\section{Methods (by generic)}{
\itemize{
\item \code{formula}: Get model formula

\item \code{formula<-}: Set formula

\item \code{n_neighbours}: Get maximum number of neighbours

\item \code{p_far_neighbours}: Get percentage of far neighbours

\item \code{distance_units}: Get distance units

\item \code{distance_units<-}: Set distance units

\item \code{max_distance}: Get maximum distance for neighbours

\item \code{obs_dag_method}: Get obs_dag_method

\item \code{extras}: Get extra settings

\item \code{extras<-}: Set miscellaneous settings
}}

\section{Slots}{

\describe{
\item{\code{formula}}{A formula}

\item{\code{n_neighbours}}{The number of parents for each node in the graph.}

\item{\code{p_far_neighbours}}{The proportion of parents randomly chosen for each node.}

\item{\code{distance_units}}{The units used for distance calculation.  Must be compatible
with units::set_units.}

\item{\code{max_distance}}{The maximum distance to look for parents. Does not affect
parents coming from p_far_neighbours.}

\item{\code{obs_dag_method}}{Which method should be used for constructing the transient graph?}

\item{\code{extras}}{Any additional settings that aren't essential for the model. E.g.
could hold an inla.mesh object.}
}}

