% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/staRVe_settings.R
\docType{class}
\name{staRVe_settings-class}
\alias{staRVe_settings-class}
\alias{staRVe_settings}
\alias{formula,staRVe_settings-method}
\alias{formula<-,staRVe_settings-method}
\alias{n_neighbours,staRVe_settings-method}
\alias{distance_units,staRVe_settings-method}
\alias{distance_units<-,staRVe_settings-method}
\alias{max_distance,staRVe_settings-method}
\alias{max_distance<-,staRVe_settings-method}
\alias{.time_name,staRVe_settings-method}
\title{An S4 class to hold extra settings for a staRVe model}
\usage{
\S4method{formula}{staRVe_settings}(x)

\S4method{formula}{staRVe_settings}(x) <- value

\S4method{n_neighbours}{staRVe_settings}(x)

\S4method{distance_units}{staRVe_settings}(x)

\S4method{distance_units}{staRVe_settings}(x) <- value

\S4method{max_distance}{staRVe_settings}(x)

\S4method{max_distance}{staRVe_settings}(x) <- value

\S4method{.time_name}{staRVe_settings}(x)
}
\arguments{
\item{x}{An object}

\item{value}{A replacement value}
}
\description{
An S4 class to hold extra settings for a staRVe model
}
\section{Methods (by generic)}{
\itemize{
\item \code{formula}: Get model formula

\item \code{formula<-}: Set model formula

\item \code{n_neighbours}: Get maximum number of neighbours used in graphs

\item \code{distance_units}: Get distance units used for the model.

\item \code{distance_units<-}: Set distance units

\item \code{max_distance}: Get maximum distance for edges in the graph

\item \code{max_distance<-}: Set maximum distance for edges in the graph

\item \code{.time_name}: Get the name of the time variable used in formula
(for internal use only)
}}

\section{Slots}{

\describe{
\item{\code{formula}}{A formula}

\item{\code{n_neighbours}}{The number of parents for each node in the graph.}

\item{\code{distance_units}}{The units used for distance calculation.  Must be compatible
with units::set_units.}

\item{\code{max_distance}}{The maximum allowable distance for edges in the transient graph.}
}}

\seealso{
Other staRVe_classes: 
\code{\link{TMB_out-class}},
\code{\link{dag-class}},
\code{\link{long_stars-class}},
\code{\link{staRVe_classes}},
\code{\link{staRVe_model-class}},
\code{\link{staRVe_model_fit-class}},
\code{\link{staRVe_observation_parameters-class}},
\code{\link{staRVe_observations-class}},
\code{\link{staRVe_parameters-class}},
\code{\link{staRVe_process-class}},
\code{\link{staRVe_process_parameters-class}},
\code{\link{staRVe_tracing-class}}
}
\concept{staRVe_classes}
