% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{staRVe_simulate,staRVe_model-method}
\alias{staRVe_simulate,staRVe_model-method}
\title{Simulate from a staRVe_model.}
\usage{
\S4method{staRVe_simulate}{staRVe_model}(model, ...)
}
\arguments{
\item{model}{A staRVe_model object.}
}
\value{
A staRve_model object with simulated random effects and observations.
}
\description{
Simulate from a staRVe_model.
}
