% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{staRVe_simulate}
\alias{staRVe_simulate}
\title{Simulate from a staRVe_model.}
\usage{
staRVe_simulate(model, ...)
}
\arguments{
\item{model}{A staRVe_model object.}

\item{conditional}{logical. If true, new observations are simulated conditional
on the random effect values in \code{random_effects(model)}.
If false, new random effects and new observations are simulated.}
}
\value{
A staRve_model object with simulated random effects and observations.
}
\description{
Simulate a new dataset from the model, with the option to simulate a new
set of random effects as well. The parameter values used in the simulations
are those set in \code{parameters(model)}. Unless new random effects are simulated,
the random effect values are those in \code{time_effects(model)} and
\code{random_effects(model)}.
}
