% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{staRVe_tracing-class}
\alias{staRVe_tracing-class}
\title{An S4 class to hold optimization tracing for a staRVe model}
\description{
An S4 class to hold optimization tracing for a staRVe model
}
