% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/show.R, R/staRVe_tracing.R
\docType{class}
\name{staRVe_tracing-class}
\alias{staRVe_tracing-class}
\alias{staRVe_tracing}
\alias{show,staRVe_tracing-method}
\alias{opt_time,staRVe_tracing-method}
\alias{hess_time,staRVe_tracing-method}
\alias{sdr_time,staRVe_tracing-method}
\alias{parameter_hessian,staRVe_tracing-method}
\alias{parameter_covariance,staRVe_tracing-method}
\alias{timing,staRVe_tracing-method}
\title{An S4 class to hold optimization tracing for a staRVe model}
\usage{
\S4method{show}{staRVe_tracing}(object)

\S4method{opt_time}{staRVe_tracing}(x)

\S4method{hess_time}{staRVe_tracing}(x)

\S4method{sdr_time}{staRVe_tracing}(x)

\S4method{parameter_hessian}{staRVe_tracing}(x)

\S4method{parameter_covariance}{staRVe_tracing}(x)

\S4method{timing}{staRVe_tracing}(x)
}
\arguments{
\item{object}{An object}

\item{x}{An object}
}
\description{
An S4 class to hold optimization tracing for a staRVe model
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Print method

\item \code{opt_time}: Get optimization time

\item \code{hess_time}: Get hessian computation time

\item \code{sdr_time}: Get standard error computation time

\item \code{parameter_hessian}: Get parameter hessian matrix

\item \code{parameter_covariance}: Get parameter covariance matrix

\item \code{timing}: Get all timing information
}}

\section{Slots}{

\describe{
\item{\code{opt_time}}{A proc_time object. Time used to compute the ML estimates.}

\item{\code{hess_time}}{A proc_time object. Time used to compute the hessian matrix.}

\item{\code{sdr_time}}{A proc_time object. Time used to compute standard errors.}

\item{\code{parameter_hessian}}{The hessian matrix for parameter estimates.}

\item{\code{parameter_covariance}}{The covariance matrix for parameter estimates.}
}}

