% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/staRVe_tracing.R
\name{staRVe_tracing}
\alias{staRVe_tracing}
\alias{initialize,staRVe_tracing-method}
\title{Create a \code{staRVe_tracing} object}
\usage{
staRVe_tracing(x, ...)

\S4method{initialize}{staRVe_tracing}(
  .Object,
  opt_time = proc.time() - proc.time(),
  hess_time = proc.time() - proc.time(),
  sdr_time = proc.time() - proc.time(),
  parameter_hessian = matrix(numeric(0), nrow = 0, ncol = 0),
  parameter_covariance = matrix(numeric(0), nrow = 0, ncol = 0)
)
}
\description{
Create a \code{staRVe_tracing} object
}
\details{
The \code{initialize} function is not mean to be used by the user,
  use \code{staRVe_tracing} instead.
}
