% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/starve.R
\name{TMB_in}
\alias{TMB_in}
\alias{TMB_in,starve-method}
\title{Convert a model object into a form suitable for TMB input.}
\usage{
TMB_in(x)

\S4method{TMB_in}{starve}(x)
}
\value{
A list with elements data, para, map, and rand to supply to TMB::MakeADFun
}
\description{
Convert a model object into a form suitable for TMB input.
}
\section{Methods (by class)}{
\itemize{
\item \code{starve}: Convert a starve object to a list for \code{TMB::MakeADFun}
}}

\keyword{internal}
