% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/TMB_out.R, R/show.R
\docType{class}
\name{TMB_out-class}
\alias{TMB_out-class}
\alias{obj,TMB_out-method}
\alias{obj<-,TMB_out-method}
\alias{opt,TMB_out-method}
\alias{opt<-,TMB_out-method}
\alias{sdr,TMB_out-method}
\alias{sdr<-,TMB_out-method}
\alias{convergence,TMB_out-method}
\alias{show,TMB_out-method}
\title{An S4 class to hold TMB-related objects}
\usage{
\S4method{obj}{TMB_out}(x)

\S4method{obj}{TMB_out}(x) <- value

\S4method{opt}{TMB_out}(x)

\S4method{opt}{TMB_out}(x) <- value

\S4method{sdr}{TMB_out}(x)

\S4method{sdr}{TMB_out}(x) <- value

\S4method{convergence}{TMB_out}(x)

\S4method{show}{TMB_out}(object)
}
\arguments{
\item{x}{An object}

\item{value}{A replacement value}

\item{object}{An object}
}
\description{
An S4 class to hold TMB-related objects
}
\section{Methods (by generic)}{
\itemize{
\item \code{obj}: Get TMB::MakeADFun object

\item \code{obj<-}: Set TMB::MakeADFun object

\item \code{opt}: Get output of optimizer

\item \code{opt<-}: Set output of optimizer

\item \code{sdr}: Get output of TMB::sdreport

\item \code{sdr<-}: Set output of TMB::sdreport

\item \code{convergence}: Get convergence message from optimizer

\item \code{show}: Print method
}}

\section{Slots}{

\describe{
\item{\code{obj}}{The output of TMB::MakeADFun.}

\item{\code{opt}}{The output of nlminb.}

\item{\code{sdr}}{The output of TMB::sdreport.}
}}

\seealso{
Other staRVe_classes: 
\code{\link{dag-class}},
\code{\link{long_stars-class}},
\code{\link{staRVe_classes}},
\code{\link{staRVe_model-class}},
\code{\link{staRVe_model_fit-class}},
\code{\link{staRVe_observation_parameters-class}},
\code{\link{staRVe_observations-class}},
\code{\link{staRVe_parameters-class}},
\code{\link{staRVe_process-class}},
\code{\link{staRVe_process_parameters-class}},
\code{\link{staRVe_settings-class}},
\code{\link{staRVe_tracing-class}}
}
\concept{staRVe_classes}
\keyword{internal}
