% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{TMB_out-class}
\alias{TMB_out-class}
\title{An S4 class to hold TMB-related objects}
\description{
An S4 class to hold TMB-related objects
}
\section{Slots}{

\describe{
\item{\code{obj}}{The output of TMB::MakeADFun.}

\item{\code{opt}}{The output of nlminb.}

\item{\code{sdr}}{The output of TMB::sdreport.}
}}

