% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/TMB_out.R, R/show.R
\docType{class}
\name{TMB_out-class}
\alias{TMB_out-class}
\alias{convergence,TMB_out-method}
\alias{show,TMB_out-method}
\title{An S4 class to hold TMB-related objects}
\usage{
\S4method{convergence}{TMB_out}(x)

\S4method{show}{TMB_out}(object)
}
\arguments{
\item{x}{An object}

\item{object}{An object}
}
\description{
An S4 class to hold TMB-related objects
}
\section{Functions}{
\itemize{
\item \code{convergence(TMB_out)}: Get convergence message from optimizer

\item \code{show(TMB_out)}: Print method

}}
\section{Slots}{

\describe{
\item{\code{obj}}{The output of TMB::MakeADFun.}

\item{\code{opt}}{The output of nlminb.}

\item{\code{sdr}}{The output of TMB::sdreport.}
}}

\seealso{
Other starve_classes: 
\code{\link{dag-class}},
\code{\link{long_stars-class}},
\code{\link{observation_parameters-class}},
\code{\link{observations-class}},
\code{\link{parameters-class}},
\code{\link{process-class}},
\code{\link{process_parameters-class}},
\code{\link{settings-class}},
\code{\link{starve-class}},
\code{\link{tracing-class}}
}
\concept{starve_classes}
\keyword{internal}
