% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/staRVe_tracing.R, R/show.R
\docType{class}
\name{tracing-class}
\alias{tracing-class}
\alias{opt_time,tracing-method}
\alias{opt_time<-,tracing-method}
\alias{hess_time,tracing-method}
\alias{hess_time<-,tracing-method}
\alias{sdr_time,tracing-method}
\alias{sdr_time<-,tracing-method}
\alias{parameter_hessian,tracing-method}
\alias{parameter_hessian<-,tracing-method}
\alias{parameter_covariance,tracing-method}
\alias{parameter_covariance<-,tracing-method}
\alias{timing,tracing-method}
\alias{show,tracing-method}
\title{An S4 class to hold optimization tracing for a starve model}
\usage{
\S4method{opt_time}{tracing}(x)

\S4method{opt_time}{tracing}(x) <- value

\S4method{hess_time}{tracing}(x)

\S4method{hess_time}{tracing}(x) <- value

\S4method{sdr_time}{tracing}(x)

\S4method{sdr_time}{tracing}(x) <- value

\S4method{parameter_hessian}{tracing}(x)

\S4method{parameter_hessian}{tracing}(x) <- value

\S4method{parameter_covariance}{tracing}(x)

\S4method{parameter_covariance}{tracing}(x) <- value

\S4method{timing}{tracing}(x)

\S4method{show}{tracing}(object)
}
\arguments{
\item{x}{An object}

\item{value}{A replacement value}

\item{object}{An object}
}
\description{
An S4 class to hold optimization tracing for a starve model
}
\section{Functions}{
\itemize{
\item \code{opt_time,tracing-method}: Get optimization time

\item \code{opt_time<-,tracing-method}: Set optimization time (for internal use only)

\item \code{hess_time,tracing-method}: Get hessian computation time

\item \code{hess_time<-,tracing-method}: Set hessian computation time (for internal use only)

\item \code{sdr_time,tracing-method}: Get standard error computation time

\item \code{sdr_time<-,tracing-method}: Set standard error computation time (for internal use only)

\item \code{parameter_hessian,tracing-method}: Get parameter estimator hessian matrix

\item \code{parameter_hessian<-,tracing-method}: Set parameter estimator hessian matrix (for internal use only)

\item \code{parameter_covariance,tracing-method}: Get parameter estimator covariance matrix

\item \code{parameter_covariance<-,tracing-method}: Set parameter estimator covariance matrix (for internal use only)

\item \code{timing,tracing-method}: Get all timing information as a list with elements
fit, hessian, and sdr.

\item \code{show,tracing-method}: Print method
}}

\section{Slots}{

\describe{
\item{\code{opt_time}}{A proc_time object. Time elapsed while computing the ML estimates.}

\item{\code{hess_time}}{A proc_time object. Time elapsed while computing the hessian matrix.}

\item{\code{sdr_time}}{A proc_time object. Time elapsed while computing standard errors.}

\item{\code{parameter_hessian}}{The hessian matrix for parameter estimates.}

\item{\code{parameter_covariance}}{The covariance matrix for parameter estimates.}
}}

\seealso{
Other starve_classes: 
\code{\link{TMB_out-class}},
\code{\link{dag-class}},
\code{\link{long_stars-class}},
\code{\link{observation_parameters-class}},
\code{\link{observations-class}},
\code{\link{parameters-class}},
\code{\link{process-class}},
\code{\link{process_parameters-class}},
\code{\link{settings-class}},
\code{\link{starve-class}},
\code{\link{starve_classes}}
}
\concept{starve_classes}
