% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{update_staRVe_model}
\alias{update_staRVe_model}
\title{Update a \code{staRVe_model} object from a model fit.}
\usage{
update_staRVe_model(x, y)
}
\description{
Update a \code{staRVe_model} object from a model fit.
}
